% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEC.R
\name{ABEC}
\alias{ABEC}
\title{Adult Blood-based EPIC Clock (ABEC)}
\usage{
ABEC(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values.
\strong{Rows must correspond to samples and columns to CpGs.}
\code{rownames} (sample IDs) and \code{colnames} (CpG probe IDs) are required.
The matrix should not contain \code{NA} values.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Predicts biological age using the Adult Blood-based EPIC Clock (ABEC).
Developed by Lee et al., this model was trained on DNA methylation (DNAm)
data from the Norwegian Mother, Father and Child Cohort Study (MoBa)
(n = 1,592, age range: 19–59 years) using the Illumina EPIC platform.
}
\details{
The function extracts the necessary CpG coefficients from the internal
\code{ABEC_Coef} dataset and applies them to the provided beta value matrix.
It relies on the \code{calculateLinearPredictor} helper function to handle
missing probes and compute the final age estimates.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
ABEC_o <- ABEC(hannum_bmiq_m)
}
\references{
Lee, Y., Haftorn, K.L., Denault, W.R.P. et al.
Blood-based epigenetic estimators of chronological age in human adults using DNA methylation data from the Illumina MethylationEPIC array.
\emph{BMC Genomics} 2020
}
