% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_scImmuAging.R
\name{ageDonor}
\alias{ageDonor}
\title{predicted age for each individual}
\usage{
ageDonor(predictRes)
}
\arguments{
\item{predictRes}{the output from scImmuAgingCalculator() function}
}
\value{
Return predicted age for each individual
}
\description{
predicted age for each individual
}
