% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_PCClocks_SystemsAge.R, R/utils.R
\name{anti.trafo}
\alias{anti.trafo}
\title{Inverse Age Transformation for Developmental Clocks}
\usage{
anti.trafo(x, adult.age = 20)

anti.trafo(x, adult.age = 20)
}
\arguments{
\item{x}{A vector of sample ages}

\item{adult.age}{The age considered to be the cutoff for adulthood}
}
\value{
A numeric vector of age predictions reverted to the chronological age scale.

transformed age prediction
}
\description{
Reverts a transformed age prediction (e.g., from a log-scale) back to
the chronological age scale, based on the Horvath (2013) method.
}
