% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bohlin_GA.R
\name{Bohlin_GA}
\alias{Bohlin_GA}
\title{Calculate the Bohlin Gestational Age (Cord Blood)}
\usage{
Bohlin_GA(beta.m)
}
\arguments{
\item{beta.m}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}
}
\value{
#' A \strong{numeric vector} containing the predicted gestational age (in weeks)
for each sample. The vector is named with the sample IDs from the \code{rownames}
of \code{beta.m}.
}
\description{
Implements the epigenetic clock for predicting gestational age (GA) using
newborn cord blood, as described by Bohlin et al. (2016).
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Bohlin_GA.out <- Bohlin_GA(hannum_bmiq_m)
}
\references{
Bohlin J, Håberg SE, Magnus P, et al.
Prediction of gestational age based on genome-wide differentially methylated regions.
\emph{Genome Biol.} 2016
}
