% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLinearPredictor.R
\name{calculateLinearPredictor}
\alias{calculateLinearPredictor}
\title{(Internal Helper) Calculate a linear predictor for an epigenetic clock}
\usage{
calculateLinearPredictor(
  beta.m,
  coef.lv,
  clock.name = "Unknown Clock",
  verbose = TRUE
)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. (CpGs as rows, Samples as cols).}

\item{coef.lv}{A list containing the clock coefficients.
\itemize{
\item \code{coef.lv[[1]]} MUST be the numeric intercept.
\item \code{coef.lv[[2]]} MUST be a named numeric vector of CpG weights.
}}

\item{clock.name}{A string, the name of the clock for print messages.}
}
\value{
A named numeric vector of predicted ages/values for each sample.
Returns NA for all samples if \code{min.perc} threshold is not met.
}
\description{
A generic helper function to calculate the weighted linear sum
(predictor) for any epigenetic clock, given a beta matrix and a
coefficient list.
}
