% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Clocks.R
\name{CTS_Clocks}
\alias{CTS_Clocks}
\title{Predict DNA methylation age using CTS clocks}
\usage{
CTS_Clocks(
  data.m,
  CTSclocks = c("Neu-In"),
  dataType = c("bulk", "sorted"),
  CTF.m = NULL,
  tissue = c("brain", "otherTissue"),
  coreNum = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data.m}{A DNAm matrix (row: CpGs, column: samples) of the samples you want to get  DNAm age predicted by a CTS clock.}

\item{CTSclocks}{A character vector of one or more clocks to apply.
(e.g., 'Neu-In', 'Hep', c('Neu-In', 'Neu-Sin', 'Brain')).}

\item{dataType}{Type of the samples ('bulk' or 'sorted').}

\item{CTF.m}{Optional cell type fraction matrix (rows: samples, columns: cell types).
Required for 'Intrinsic' bulk clocks if tissue is not 'brain'.}

\item{tissue}{What tissue are your samples from ('brain' or 'otherTissue').}

\item{coreNum}{Number of cores for parallel computation in preprocessing.}

\item{verbose}{Logical. If \code{TRUE} (default), the function will print
progress messages to the console.}
}
\value{
A data.frame of predicted DNAm ages (samples x clocks).
}
\description{
This is the function for computing DNA methylation age using CTS (cell type
specific) clocks. The inputs including DNAm matrix, the CTS clock you want to
use, is your DNAm data from bulk tissue sample or sorted cell sample, cell
type fraction matrix if you want to use Neu-In/Glia-In/Brain clock, tissue of
your DNAm data samples and the number of cores if you want to do parallel computing.
}
\details{
This function supports a variety of Cell-Type-Specific (CTS) clocks.

\strong{Available \code{CTSclocks} include:}
\itemize{
\item \code{'Neu-In'}
\item \code{'Glia-In'}
\item \code{'Brain'}
\item \code{'Neu-Sin'}
\item \code{'Glia-Sin'}
\item \code{'Hep'}
\item \code{'Liver'}
'
The clocks are grouped below based on their biological target:
}

\strong{1. Cell-Type Specific Clocks}

These clocks are trained to measure aging in specific cell populations.
\itemize{
\item \code{'Neu-In'} (Intrinsic): Measures cell-intrinsic aging of \strong{Neurons}.
(Uses processed data: residuals/Z-scores).
\item \code{'Glia-In'} (Intrinsic): Measures cell-intrinsic aging of \strong{Glial cells}.
(Uses processed data: residuals/Z-scores).
\item \code{'Neu-Sin'} (Semi-intrinsic): Measures aging of \strong{Neurons} using raw data.
(Reflects both intrinsic aging and composition changes).
\item \code{'Glia-Sin'} (Semi-intrinsic): Measures aging of \strong{Glial cells} using raw data.
(Reflects both intrinsic aging and composition changes).
\item \code{'Hep'} (Semi-intrinsic): Measures aging of \strong{Hepatocytes} using raw data.
}

\strong{2. Non Cell-Type Specific Clocks}
\itemize{
\item \code{'Brain'} (Intrinsic): A intrinsic clock for \strong{whole brain tissue}.
(Uses processed data: residuals/Z-scores).
\item \code{'Liver'} (Semi-intrinsic): A intrinsic clock for \strong{whole liver tissue}.
(Uses raw data).
}
}
\examples{
download_OmniAgeR_example("CTS_MurphyGSE88890")
load_OmniAgeR_example("CTS_MurphyGSE88890")
agePred_df <- CTS_Clocks(Murphy_beta_m, CTSclock = c('Neu-In','Neu-Sin'), dataType = 'bulk', CTF.m = NULL, tissue = 'brain',coreNum = 1)

download_OmniAgeR_example("CTS_PaiGSE112179")
load_OmniAgeR_example("CTS_PaiGSE112179")
agePred_df <- CTS_Clocks(Pai_beta_m, CTSclock = c('Neu-In','Neu-Sin'), dataType = 'sorted', CTF.m = NULL, tissue = 'brain',coreNum = 1)

download_OmniAgeR_example("CTS_ExampleData_Liver")
load_OmniAgeR_example("CTS_ExampleData_Liver")
agePred_df <- CTS_Clocks(Liver_beta_m, CTSclock = c('Hep','Liver'), dataType = 'bulk', CTF.m = NULL, tissue = 'otherTissue',coreNum = 1)


}
\references{
Tong H, Guo X, Jacques M, Luo Q, Eynon N, Teschendorff AE.
Cell-type specific epigenetic clocks to quantify biological age at cell-type resolution.
\emph{Aging} 2024
}
