% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_CTF_Clock.R
\name{DNAm_CTF_Clock}
\alias{DNAm_CTF_Clock}
\title{DNA Methylation Cell-Type Fraction (CTF) Aging Clock}
\usage{
DNAm_CTF_Clock(CTF_m)
}
\arguments{
\item{CTF_m}{A numeric matrix or data frame where rows are samples and columns are cell types.
Must contain the specific cell types required by the model (e.g., predicted by EpiDISH).}
}
\value{
A named numeric vector of predicted ages.
}
\description{
Predicts biological age based on immune cell type fractions derived from DNA methylation data.
}
\examples{

download_OmniAgeR_example("TZH_example_CTF")
load_OmniAgeR_example("TZH_example_CTF")
DNAm_CTF_Clock_o<-DNAm_CTF_Clock(CTF_m = TZH_Frac_m)

}
