% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnamExample}
\alias{dnamExample}
\title{Example DNA methylation data}
\format{
This resource is a \code{list} of length 2, containing:
\describe{
\item{betaM}{A numeric matrix of BMIQ-normalized DNA methylation beta values.}
\item{phenotype}{A \code{data.frame} of 6 phenotypic variables associated with the samples: \code{Age}, \code{Source}, \code{Plate}, \code{Sex}, \code{Ethn} (Ethnicity), and \code{SmkStat} (Smoking Status).}
}
}
\source{
The original data is derived from the Hannum et al. study (GEO Accession: GSE40279).
Preprocessing pipelines reference: Luo, Q. et al. \emph{Genome Med} 15, 59 (2023).

\strong{License:}
The original data was deposited in GEO as a public resource. de
}
\usage{
data(dnamExample)
}
\description{
A lightweight example dataset containing whole blood DNA methylation profiles
and matched phenotypic metadata from a random subset of the Hannum cohort.
It is provided to test DNAm aging clocks in the \code{OmniAgeR} package.
}
\details{
\strong{How it was generated:}
The original Hannum cohort dataset (Illumina 450k platform) was downloaded and
preprocessed as previously described by Luo et al. (\emph{Genome Med}, 2023).
Following BMIQ normalization to correct for probe-type bias, I randomly extracted
a subset of 3 samples to create a lightweight testing object. Subsequently, I
retained only the CpGs commonly utilized by the majority of epigenetic clocks.
The beta matrix and clinical variables were then bundled into a standard R list
to seamlessly interface with the \code{OmniAgeR} pipeline.
}
\keyword{datasets}
