% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalAgingScore.R
\name{.enforceDependencies}
\alias{.enforceDependencies}
\title{Enforce Clock Dependencies}
\usage{
.enforceDependencies(clocksToRun)
}
\arguments{
\item{clocksToRun}{A character vector of requested individual clocks.}
}
\value{
A character vector including necessary dependent clocks.
}
\description{
Automatically adds precursor clocks to the execution list if a downstream
clock (e.g., DNAmFitAge) depends on their results.
}
\keyword{internal}
