% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFeature.R
\name{.formatLinearWeights}
\alias{.formatLinearWeights}
\title{Standardize Linear Clock Weights Format}
\usage{
.formatLinearWeights(rawCoef)
}
\arguments{
\item{rawCoef}{A data.frame containing raw clock coefficients.}
}
\value{
A standardized data.frame.
}
\description{
Internal helper to ensure all linear clock weights are returned as a
standard data.frame with 'feature' and 'coef' columns.
}
\keyword{internal}
