% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_OmniAgeR_example}
\alias{download_OmniAgeR_example}
\title{Download OmniAgeR Example Data}
\usage{
download_OmniAgeR_example(
  datasets = c("Hannum_example.rda", "GA_example.rda", "LungInv.rda",
    "Tursiops_example.rda", "TZH_example_CTF.rda",
    "brain_frohlich_control_example_15donors.rda", "seu_gabitto_2024_filtered",
    "Yazar_CD4T_CD8T_example.rda", "CTS_ExampleData_Liver.rda", "CTS_MurphyGSE88890.rda",
    "CTS_PaiGSE112179.rda"),
  path = NULL,
  force = FALSE,
  ...
)
}
\arguments{
\item{datasets}{A character vector of datasets to download (e.g., "seu_gabitto_2024_filtered").}

\item{path}{Path to save data. Defaults to \code{get_OmniAgeR_path()}.}

\item{force}{Logical. If TRUE, overwrites existing files.}
}
\description{
Downloads large example datasets (.rda) required for vignettes and examples.
Data is stored in the package's user data directory.
}
