% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalAgingScore.R
\name{EpiAge}
\alias{EpiAge}
\title{Estimate epigenetic age}
\usage{
EpiAge(
  data.m,
  clock_names = "all",
  ages.v = NULL,
  sex.v = NULL,
  PCClocks_RData = NULL,
  SystemsAge_RData = NULL,
  Stoch_refM.m = NULL,
  CTS_dataType = NULL,
  CTS_CTF.m = NULL,
  CTS_tissue = NULL,
  CTS_coreNum = NULL,
  CTF_m
)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}

\item{clock_names}{A character vector specifying which clocks to calculate.
This can include individual clock names (e.g., "Horvath2013") or special
keywords:
\itemize{
\item \code{"all"}: (Default) Calculates all available clocks:
\code{cellular_aging} + \code{chronological} + \code{biological} + \code{causal}+ \code{stochastic} + \code{ensemble} + \code{cross_species} + \code{celltype_specific}.
\item \code{"cellular_aging"}: Calculates all cellular aging clocks (both \code{mitotic} and \code{dnamtl}).
\item \code{"mitotic"}: Calculates only the mitotic clocks (e.g., \code{epiTOC1}, \code{HypoClock}).
\item \code{"dnamtl"}: Calculates only the DNAm Telomere Length clock (\code{DNAmTL},\code{PCDNAmTL}).
\item \code{"mitotic"}: Calculates all defined mitotic clocks (e.g., \code{epiTOC1}, \code{epiTOC2}, \code{HypoClock}).
\item \code{"chronological"}: Calculates chronological age clocks (e.g., \code{Horvath2013}, \code{Hannum}, \code{ZhangClock}).
\item \code{"biological"}: Calculates biological age clocks (e.g., \code{PhenoAge}, \code{DunedinPACE}, \code{GrimAge1}, \code{GrimAge2}).
\item \code{"causal"}: Calculates causality-related clocks (\code{CausalClock}).
\item \code{"stochastic"}: Calculates the stochastic clocks (StocH, StocP, StocZ) as a single bundle.
\item \code{"celltype_specific"}: Calculates all cell-type-specific clocks (\code{Neu-In}, \code{Glia-Sin}, etc.) as a single bundle.
}

Multiple keywords and individual names can be combined (e.g., \code{c("Horvath2013", "GrimAge2")}).
Use \code{listEpiAge()} to see a categorized list of all available clocks.}

\item{ages.v}{Optional argument representing the chronological ages or surrogates thereof of the samples. Vector must be of same length as the number of samples.}

\item{sex.v}{Optional argument representing the sex of the samples. Vector must be of same length as the number of samples.}

\item{PCClocks_RData}{\strong{Required if any PC Clock is requested.} This argument specifies the
PC clock model data (the \code{PCClocks_data.qs2} file). This can be:
\itemize{
\item A \code{character} string: The path to the \emph{directory} containing
the \code{PCClocks_data.qs2} file.
\item A \code{list}: The pre-loaded data object (from
\code{load_OmniAgeR_data(object_name = "PCClocks_data")}).
}
See \code{?PCClocks} for details.}

\item{SystemsAge_RData}{\strong{Required if 'SystemsAge' is requested.} This argument specifies the
SystemsAge model data (the \code{SystemsAge_data.qs} file). This can be:
\itemize{
\item A \code{character} string: The path to the \emph{directory} containing
the \code{SystemsAge_data.qs} file.
\item A \code{list}: The pre-loaded data object (from
\code{load_OmniAgeR_data(object_name = "SystemsAge_data")}).
}
See \code{?SystemsAge} for details.}

\item{Stoch_refM.m}{\strong{Optional (for Stochastic Clocks).} A reference matrix for cell-type
deconvolution (e.g., from \code{EpiDISH}). If provided (along with \code{ages.v}),
this will be used to calculate Intrinsic Age Acceleration (IAA). (Default: \code{NULL}).}

\item{CTS_dataType}{\strong{Required if any Cell-Type-Specific Clock is requested.}
Character string. The type of input sample for CTS_Clocks: \code{'bulk'} or \code{'sorted'}.}

\item{CTS_CTF.m}{\strong{Optional (for CTS_Clocks).} Cell type fraction matrix (Samples x CellTypes).
See \code{?CTS_Clocks} for details on when this is required.}

\item{CTS_tissue}{\strong{Required if any Cell-Type-Specific Clock is requested.}
Character string. The source tissue for CTS_Clocks: \code{'brain'} or \code{'otherTissue'}.}

\item{CTS_coreNum}{\strong{Optional (for CTS_Clocks).} Number of cores for parallel computation
in \code{CTS_Clocks} preprocessing.}
}
\value{
A list containing the entries for the calculated clocks.
\itemize{
\item \code{epiTOC1}: The epiTOC1 score.
\item \code{epiTOC2}: A list (tnsc, tnsc2, irS, etc.).
\item \code{epiTOC3}: A list (tnsc, tnsc2, irS, etc.).
\item \code{stemTOCvitro}: The stemTOCvitro score.
\item \code{stemTOC}: The stemTOC score.
\item \code{RepliTali}: The RepliTali score.
\item \code{HypoClock}: The HypoClock score.
\item \code{EpiCMIT_hyper}: The EpiCMIT_hyper score.
\item \code{EpiCMIT_hypo}: The EpiCMIT_hypo score.
\item \code{DNAmTL}: The Leukocyte telomere length.
\item \code{Horvath2013}: Horvath epigenetic clock age.
\item \code{Hannum}: Hannum epigenetic clock age.
\item \code{Lin}: Lin epigenetic clock age.
\item \code{VidalBralo}: VidalBralo epigenetic clock age.
\item \code{ZhangClock}: Zhang epigenetic clock age.
\item \code{Horvath2018}: Horvath skin & blood clock age.
\item \code{Bernabeu_cAge}: Bernabeu cAge.
\item \code{CorticalClock}: CorticalClock age.
\item \code{PedBE}: PedBE clock age.
\item \code{Zhang10}: Zhang 10-CpG score.
\item \code{PhenoAge}: PhenoAge score.
\item \code{DunedinPACE}: DunedinPACE score.
\item \code{GrimAge1}: A data.frame (Sample, Age, Female, DNAm..., DNAmGrimAge1)
\item \code{GrimAge2}: A data.frame (Sample, Age, Female, DNAm..., DNAmGrimAge2).
\item \code{CausalClock}: A list (Causal, Damage, Adaptation scores).
\item \code{IC_Clock}: The Intrinsic Capacity (IC) score
\item \code{ABEC}: ABEC age.
\item \code{eABEC}: eABEC age.
\item \code{cABEC}: cABEC age
\item \code{PipekElasticNet}: PipekElasticNet age
\item \code{PipekFilteredh}:  PipekFilteredh age
\item \code{PipekRetrainedh}:  PipekRetrainedh age
\item \code{WuClock}:  WuClock age
\item \code{PCClocks}: (If requested) A data.frame containing the original \code{Sample_ID}, \code{Age}, and \code{Female} columns, appended with 14 new columns for the calculated PC clock values (including \code{PCHorvath2013}, \code{PCPhenoAge}, \code{PCDNAmTL}, etc.).
\item \code{SystemsAge}: (If requested) A data.frame where the first column is \code{Sample_ID} and the subsequent 13 columns contain the calculated scores
\item \code{DNAmFitAge}: (If requested) A single data.frame containing DNAmFitAge, FitAgeAccel, and all 6 related fitness biomarkers.
\item \code{StochClocks}: (If requested) A list containing the \code{mage} (DNAm ages)  and \code{eaa} (Extrinsic Age Acceleration, if \code{ages.v} was provided).Note: Intrinsic Age Acceleration (IAA) is not calculated via the \code{EpiAge} wrapper. Please call \code{StochClocks()} directly if IAA is needed.
}
}
\description{
Estimates epigenetic age by applying a suite of established DNA methylation
(DNAm) aging clocks to a DNAm matrix. This wrapper function provides a
unified interface to run multiple clocks, such as "Horvath2013" or
"PhenoAge", either individually or as predefined groups (e.g., "cellular_aging").
}
\details{
The EpiAge function implements a wide range of epigenetic clocks.
This includes cellular aging clocks (e.g., \code{epiTOC2}, \code{DNAmTL}), chronological age clocks (e.g., \code{Horvath2013}, \code{Hannum}, \code{ZhangClock}),
biological age clocks (e.g., \code{PhenoAge}, \code{DunedinPACE}, \code{GrimAge2}), causality-related clocks ( \code{CausalAge}, \code{DamAge}, \code{AdaptAge}),
Stochastic Clocks(\code{StocH}, \code{StocP}, \code{StocZ}), and Cell-Type Specific Clocks(e.g., \code{Neu-In}, \code{Glia-In}).

=== Cellular Aging Clocks ===
This category includes clocks measuring cell proliferation (mitotic clocks) and telomere length.
\strong{Mitotic Clocks}
\itemize{
\item \code{epiTOC1}: (Yang et al. 2016) Average beta value of 385 promoter CpGs.
\item \code{epiTOC2}: (Teschendorff et al. 2020) Estimates stem cell divisions using a dynamic model.
\item \code{epiTOC3}: Estimates stem cell divisions based on unmethylated population doubling associated CpGs.
\item \code{stemTOCvitro}: (Yang et al. 2016) The 0.95 upper quantile of the 629 stemTOCvitro CpGs. The stemTOCvitro CpGs are promoters CpGs that are unmethylated in fetal tissue-types and undergo DNA hypermethylation with increased population-doublings.
\item \code{stemTOC}: (Yang et al. 2016) The 0.95 upper quantile of the 371 stemTOC CpGs. Compared to stemTOCvitro CpGs, the stemTOC CpGs are filtered for significant DNA hypermethylation with chronological age in large in-vivo datasets.
\item \code{RepliTali}: (Endicott et al. 2022) Based on 87 population doubling associated hypomethylated CpGs.
\item \code{HypoClock}: (Yang et al. 2016) Based on hypomethylation at 678 solo-WCGW sites.
\item \code{EpiCMIT_hyper}: (Duran-Ferrer et al. 2020) Average beta value of 184 age-associated hypermethylated CpGs.
\item \code{EpiCMIT_hypo}: (Duran-Ferrer et al. 2020) Average beta value of 1164 age-associated hypomethylated CpGs.
\item \code{EpiCMIT_hypo}: (Duran-Ferrer et al. 2020) Average beta value of 1164 age-associated hypomethylated CpGs.
}

\strong{DNAm Telomere Length (TL) Clocks}
\itemize{
\item \code{DNAmTL}: (Lu AT et al. 2019) Calculating the Leukocyte telomere length.
\item \code{PCDNAmTL}: (Lu AT et al. 2019) (Higgins-Chen et al. 2022)
Computationally-bolstered versions of the original clocks. These "PC clocks"
are retrained using principal components (PCs) derived from a large
CpG set to minimize technical noise and improve reliability
}

=== Chronological Clocks ===
\itemize{
\item \code{Horvath2013}: (Horvath 2013) The original pan-tissue clock from 353 CpGs.
\item \code{Hannum}: (Hannum et al. 2013) Blood-specific clock from 71 CpGs.
\item \code{Lin}: (Lin et al. 2016) Clock based on 99 CpGs.
\item \code{VidalBralo}: (Vidal-Bralo et al. 2016) Clock based on 8 CpGs.
\item \code{ZhangClock}: (Zhang et al. 2019) Improved-precision clock from 514 CpGs.
\item \code{Horvath2018}: (Horvath et al. 2018) Skin & blood clock.
\item \code{Bernabeu_cAge}: (Bernabeu et al. 2022) Clock based on 3225 CpGs.
\item \code{CorticalClock}: (Shireby et al. 2020) Crain cortical clock based on 347 CpGs.
\item \code{PedBE}: (McEwen et al. 2019) Pediatric buccal epithelial clock in Children.
\item \code{CentenarianClock}: (Eric Dec et al. 2023) Centenarian Epigenetic Clocks.
\item \code{Retro_age}: (Ndhlovu et al. 2024) Retroelement-based clock developed on the EPIC v1/v2 arrays.
\item \code{ABEC}: (Lee et al. 2020) Adult Blood-based EPIC Clock.
\item \code{eABEC}: (Lee et al. 2020) Extended Adult Blood-based EPIC Clock.
\item \code{cABEC}: (Lee et al. 2020) Common Adult Blood-based EPIC Clock.
\item \code{PipekElasticNet}: (Pipek et al. 2023) Pipek's Multi-tissue Elastic Net Epigenetic Clock (239 CpGs).
\item \code{PipekFilteredh}: (Pipek et al. 2023) Pipek's Filtered Horvath Epigenetic Clock (272 CpGs).
\item \code{PipekRetrainedh}: (Pipek et al. 2023) Pipek's Retrained Horvath Epigenetic Clock (308 CpGs).
\item \code{WuClock}: (Wu et al. 2019) Wu's Epigenetic Clock for Pediatric Age Estimation.
\item \code{PCHorvath2013}, \code{PCHorvath2018}, \code{PCHannum}: (Higgins-Chen et al. 2022)
Computationally-bolstered versions of the original clocks. These "PC clocks"
are retrained using principal components (PCs) derived from a large
CpG set to minimize technical noise and improve reliability
}

=== Biological Clocks ===
\itemize{
\item \code{Zhang10}: (Zhang et al. 2017) 10-CpG clock associated with mortality.
\item \code{PhenoAge}: (Levine et al. 2018) Predicts phenotypic age from 513 CpGs.
\item \code{DunedinPACE}: (Belsky et al. 2022) Quantifies the pace of biological aging.
\item \code{GrimAge1}: The original GrimAge clock (Lu et al. 2019).
\item \code{GrimAge2}: (Lu et al. 2022) Updated composite biomarker of mortality risk.
\item \code{PCPhenoAge}, \code{PCGrimAge1}: (Higgins-Chen et al. 2022)
Computationally-bolstered PC versions of the PhenoAge and GrimAge1
clocks, retrained on principal components for enhanced reliability.
\item \code{DNAmFitAge}: (McGreevy et al. 2023) Biological age indicator incorporating DNAmGrimAge and 3 DNAm-based physical fitness markers.
\item \code{IC_Clock}: (Fuentealba et al. 2025) The Intrinsic Capacity (IC) Clock based on 91 CpGs.
\item \code{SystemsAge}: (Sehgal et al. 2025) The Systems Age and 11 System-Specific Scores.
}

=== Causal Clocks ===
\itemize{
\item \code{CausalAge}, \code{DamAge}, \code{AdaptAge}: (Ying et al. 2024) Three clocks derived from a causality-enriched model. They dissect aging into distinct components: 'Causal' (CausalAge), 'Damage' (DamAge), and Adaptation' (AdaptAge).
}

=== Stochastic Clocks ===
\itemize{
\item \code{StocH}, \code{StocP}, \code{StocZ}: (Tong et al. 2024) Stochastic analogues of the
Horvath, PhenoAge, and Zhang clocks, trained on artificial cohorts to
quantify the stochastic component of epigenetic aging.
}

=== Cell-Type Specific Clocks ===
\itemize{
\item \code{Neu-In}, \code{Glia-In}, \code{Brain}: (Tong et al. 2024) Intrinsic clocks for neurons, glia, and whole brain.
\item \code{Neu-Sin}, \code{Glia-Sin}, \code{Hep}, \code{Liver}: (Tong et al. 2024) Semi-intrinsic clocks for neurons, glia, hepatocytes, and whole liver.
}
}
\examples{
download_OmniAgeR_example("LungInv")
load_OmniAgeR_example("LungInv")
EpiAge.o<-EpiAge(data.m = bmiq.m,clock_names="mitotic",df$Age)

download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
age <- PhenoTypesHannum_lv$Age
sex <- ifelse(PhenoTypesHannum_lv$Sex=="F","Female","Male")
PCClocks_RData <- load_OmniAgeR_data(object_name = "PCClocks_data")
SystemsAge_RData <- load_OmniAgeR_data(object_name = "SystemsAge_data")

clock_names<-  c("cellular_aging", "chronological", "biological", "causal", "stochastic")

EpiAge.o <- EpiAge(data.m = hannum_bmiq_m,clock_names = clock_names, ages.v = age, sex.v = sex,PCClocks_RData=PCClocks_RData,SystemsAge_RData=SystemsAge_RData)

}
\references{
Yang Z, Wong A, Kuh D, et al.
Correlation of an epigenetic mitotic clock with cancer risk.
\emph{Genome Biol.} 2016

Teschendorff AE.
A comparison of epigenetic mitotic-like clocks for cancer risk prediction.
\emph{Genome Med.} 2020

Endicott JL, Nolte PA, Shen H, Laird PW.
Cell division drives DNA methylation loss in late-replicating domains in primary human cells.
\emph{Nat Commun.} 2022

Duran-Ferrer M, Clot G, Nadeu F, et al.
The proliferative history shapes the DNA methylome of B-cell tumors and predicts clinical outcome.
\emph{Nat Cancer} 2020
Horvath S.
DNA methylation age of human tissues and cell types.
\emph{Genome Biol.} 2013

Zhang Q, Vallerga CL, Walker RM, et al.
Improved precision of epigenetic clock estimates across tissues and its implication for biological ageing.
\emph{Genome Med.} 2019

Levine ME, Lu AT, Quach A, et al.
An epigenetic biomarker of aging for lifespan and healthspan.
\emph{Aging} 2018

Belsky DW, Caspi A, Corcoran DL, et al.
DunedinPACE, a DNA methylation biomarker of the pace of aging.
\emph{eLife} 2022

Lu AT, Quach A, Wilson JG, et al.
DNA methylation GrimAge strongly predicts lifespan and healthspan
\emph{Aging} 2019

Lu AT, Binder AM, Zhang J, et al.
DNA methylation GrimAge version 2.
\emph{Aging} 2022

Ying K, Liu H, Tarkhov AE, et al.
Causality-enriched epigenetic age uncouples damage and adaptation.
\emph{Nat Aging} 2024

Hannum G, Guinney J, Zhao L, et al.
Genome-wide methylation profiles reveal quantitative views of human aging rates.
\emph{Mol Cell.} 2013

Lin Q, Weidner CI, Costa IG, et al.
DNA methylation levels at individual age-associated CpG sites can be indicative for life expectancy.
\emph{Aging} 2016

Vidal-Bralo L, Lopez-Golan Y, Gonzalez A.
Simplified Assay for Epigenetic Age Estimation in Whole Blood of Adults.
\emph{Front Genet.} 2016

Zhang Y, Wilson R, Heiss J, et al.
DNA methylation signatures in peripheral blood strongly predict all-cause mortality.
\emph{Nat Commun.} 2017

Horvath S, Oshima J, Martin GM, et al.
Epigenetic clock for skin and blood cells applied to Hutchinson Gilford Progeria Syndrome and ex vivo studies.
\emph{Aging} 2018

McEwen LM, O'Donnell KJ, McGill MG, et al.
The PedBE clock accurately estimates DNA methylation age in pediatric buccal cells.
\emph{Proc Natl Acad Sci U S A.} 2020

Dec, E., Clement, J., Cheng, K. et al.
Centenarian clocks: epigenetic clocks for validating claims of exceptional longevity.
\emph{GeroScience} 2023

Ndhlovu LC, Bendall ML, Dwaraka V, et al.
Retro-age: A unique epigenetic biomarker of aging captured by DNA methylation states of retroelements.
\emph{Aging Cell.} 2024

Higgins-Chen AT, Thrush KL, Wang Y, et al.
A computational solution for bolstering reliability of epigenetic clocks: Implications for clinical trials and longitudinal tracking.
\emph{Nat Aging.} (2022).

Sehgal, R., Markov, Y., Qin, C. et al.
Systems Age: a single blood methylation test to quantify aging heterogeneity across 11 physiological systems.
\emph{Nat Aging} (2025).

Tong, H., Dwaraka, V.B., Chen, Q. et al.
Quantifying the stochastic component of epigenetic aging.
\emph{Nat Aging} (2024). \doi{10.1038/s43587-024-00636-6}

Fuentealba M, Rouch L, Guyonnet S, et al.
A blood-based epigenetic clock for intrinsic capacity predicts mortality and is associated with clinical, immunological and lifestyle factors.
\emph{Nature Aging.} 2025

McGreevy KM, Radak Z, Torma F, et al.
DNAmFitAge: biological age indicator incorporating physical fitness.
\emph{Aging} 2023

Lu AT, Seeboth A, Tsai PC, et al.
DNA methylation-based estimator of telomere length
\emph{Aging} 2019

Tong H, Guo X, Jacques M, Luo Q, Eynon N, Teschendorff AE.
Cell-type specific epigenetic clocks to quantify biological age at cell-type resolution.
\emph{Aging} 2024

Lee, Y., Haftorn, K.L., Denault, W.R.P. et al.
Blood-based epigenetic estimators of chronological age in human adults using DNA methylation data from the Illumina MethylationEPIC array.
\emph{BMC Genomics} 2020

Pipek, O.A., Csabai, I.
A revised multi-tissue, multi-platform epigenetic clock model for methylation array data.
\emph{J Math Chem} 2023

Wu, Xiaohui et al.
DNA methylation profile is a quantitative measure of biological aging in children
\emph{Aging} 2019
}
\seealso{
\code{listEpiAge()}

\code{PCClocks()},
\code{SystemsAge()},
\code{DNAmFitAge()},
\code{StochClocks()}

\code{epiTOC1()}, \code{epiTOC2()}, \code{epiTOC3()}, \code{RepliTali()}, \code{HypoClock()},
\code{stemTOCvitro()}, \code{stemTOC()}, \code{EpiCMIT()},\code{DNAmTL()},
\code{Horvath2013()}, \code{Hannum()}, \code{Lin()}, \code{VidalBralo()}, \code{ZhangClock()},
\code{Horvath2018()}, \code{Bernabeu_cAge()}, \code{CorticalClock()}, \code{PedBE()},
\code{CentenarianClock()}, \code{Retro_age()},
\code{Zhang10()}, \code{PhenoAge()}, \code{DunedinPACE()}, \code{GrimAge1()}, \code{GrimAge2()},
\code{IC_Clock()}, \code{CausalClock()}, \code{CausalClock()},\code{ABEC()},\code{eABEC()},\code{cABEC()},
\code{PipekElasticNet()},\code{PipekFilteredh()},\code{PipekRetrainedh()},\code{WuClock()}.
}
