% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_GA.R
\name{EPIC_GA}
\alias{EPIC_GA}
\title{The Gestational Age (GA) clock based on 176 Illumina EPIC CpGs}
\usage{
EPIC_GA(beta.m)
}
\arguments{
\item{beta.m}{A matrix of beta values (CpGs in rows, samples in columns).  This matrix must be pre-normalized (e.g., via BMIQ) and imputed.}
}
\value{
A named vector of predicted Gestational Ages (in weeks).
}
\description{
The Gestational Age (GA) clock based on 176 Illumina EPIC CpGs
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
EPIC_GA.out <- EPIC_GA(hannum_bmiq_m)
}
\references{
Haftorn KL, Lee Y, Denault WRP, et al.
An EPIC predictor of gestational age and its application to newborns conceived by assisted reproductive technologies.
\emph{Clin Epigenetics.} 2021
}
