% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalAgingScore.R
\name{EpiGA}
\alias{EpiGA}
\title{Estimate Gestational Age (GA)}
\usage{
EpiGA(data.m, clock_names = "all")
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs
and columns labeling samples.}

\item{clock_names}{A character vector specifying which GA clocks to calculate.
Defaults to \code{"all"}.
\itemize{
\item \code{"all"}: Calculates all available GA clocks.
\item \code{c("Knight_GA", "Lee_GA", ...)}: Calculates only the specified clocks.
}

Use \code{listEpiGA()} to see a character vector of all available GA clock names.}
}
\value{
A list containing the entries for the calculated clocks.
\itemize{
\item \code{Bohlin_GA}: The Bohlin Gestational Age.
\item \code{EPIC_GA}: The EPIC Gestational Age.
\item \code{Lee_GA}: A list (LeeControl, LeeRobust, LeeRefinedRobust).
\item \code{Knight_GA}: he Knight Gestational Age
\item \code{Mayne_GA}: The Mayne Placental Gestational Age.
}

Each vector is named with the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Estimates epigenetic Gestational Age (GA) by applying one or more
established GA clocks to a DNA methylation (DNAm) matrix.
\strong{All calculated estimates are returned in units of gestational weeks.}
}
\details{
This wrapper function provides a unified interface to run multiple
GA clocks. It processes the \code{clock_names} argument and calls the
appropriate underlying clock functions.
}
\examples{
data("GA_example")
download_OmniAgeR_example("GA_example")
load_OmniAgeR_example("GA_example")
EpiGA.o <- EpiGA(GA_m)


}
\references{
Bohlin J, Håberg SE, Magnus P, et al.
Prediction of gestational age based on genome-wide differentially methylated regions.
\emph{Genome Biol.} 2016

Haftorn KL, Lee Y, Denault WRP, et al.
An EPIC predictor of gestational age and its application to newborns conceived by assisted reproductive technologies.
\emph{Clin Epigenetics.} 2021

Knight AK, Craig JM, Theda C, et al.
An epigenetic clock for gestational age at birth based on blood methylation data.
\emph{Genome Biol.} 2016

Lee Y, Choufani S, Weksberg R, et al.
Placental epigenetic clocks: estimating gestational age using placental DNA methylation levels.
\emph{Aging} 2019

Mayne BT, Leemaqz SY, Smith AK, Breen J, Roberts CT, Bianco-Miotto T.
Accelerated placental aging in early onset preeclampsia pregnancies identified by DNA methylation.
\emph{Epigenomics} 2017
}
\seealso{
\code{\link[=listEpiGA]{listEpiGA()}}
\code{Bohlin_GA()}, \code{EPIC_GA()}, \code{Knight_GA()}, \code{Lee_GA()}, \code{Mayne_GA()}

\code{\link[=listEpiGA]{listEpiGA()}}
\code{Bohlin_GA()}, \code{EPIC_GA()}, \code{Knight_GA()}, \code{Lee_GA()}, \code{Mayne_GA()}
}
