% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTOC1.R
\name{epiTOC1}
\alias{epiTOC1}
\title{Estimate epiTOC1 score}
\usage{
epiTOC1(x, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{x}{A numeric matrix, \code{data.frame}, or \code{SummarizedExperiment}
object containing DNA methylation beta values. Rows should be CpG probes and
columns individual samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
The epiTOC1 score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and
will return the epiTOC1 score.
}
\details{
The epiTOC1 score is calculated as the average beta value of 385 CpGs
described in Yang et al. 2014. These CpGs are promoter CpG sites that
localize to Polycomb group target genes that are unmethylated in 11
different fetal tissue types.
}
\examples{
data(dnamExample)
beta_matrix <- dnamExample[[1]]

# Example 1: Direct Matrix Input
epiTOC1Out  <- epiTOC1(x = beta_matrix, verbose = FALSE)

# Example 2: SummarizedExperiment Input
\dontrun{
library(SummarizedExperiment)
pheno_data <- dnamExample[[2]]
rownames(pheno_data) <- colnames(beta_matrix)

se_obj <- SummarizedExperiment(
      assays = list(beta = beta_matrix),
      colData = pheno_data
    )
    
epiTOC1Out <- epiTOC1(x = se_obj, verbose = FALSE)
}
}
\references{
Yang Z, Wong A, Kuh D, et al.
Correlation of an epigenetic mitotic clock with cancer risk.
\emph{Genome Biol.} 2016
}
