% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exact_CpGs.R
\name{getClockProbes}
\alias{getClockProbes}
\title{Get CpG Probes from Various DNAm Biomarkers}
\usage{
getClockProbes(
  clock_names = "all",
  PCClocks_RData = NULL,
  SystemsAge_RData = NULL
)
}
\arguments{
\item{clock_names}{A character vector specifying which models to retrieve probes for.
Options include specific model names (e.g., \code{"GrimAge2"}, \code{"McCartney_Trait"}) or keywords for groups:
\itemize{
\item \code{"all"}: (Default) Retrieve probes for all available models.
\item \code{"chronological"}: Chronological age predictors (First-generation clocks, e.g., Horvath, Hannum).
\item \code{"biological"}: Biological age predictors focused on mortality/morbidity (e.g., PhenoAge, GrimAge, DunedinPACE).
\item \code{"cellular_aging"}: A bundle including both mitotic clocks and DNAm telomere length estimators.
\item \code{"mitotic"}: Estimates of cumulative cell division rates (e.g., epiTOC, HypoClock).
\item \code{"dnamtl"}: DNA methylation-based estimators of telomere length (e.g., DNAmTL).
\item \code{"causal"}: Clocks trained using causal inference methods to distinguish damage from adaptation (e.g., DamAge, AdaptAge).
\item \code{"stochastic"}: Stochastic analogues of the Horvath, PhenoAge, and Zhang clocks (e.g., StocH, StocZ).
\item \code{"cross_species"}: Clocks applicable across species or specific to mouse (e.g., UniversalPanMammalian).
\item \code{"gestational"}: Predictors of gestational age (GA) (e.g., Bohlin_GA, Knight_GA, EPIC_GA).
\item \code{"celltype_specific"}: Cell-type specific clocks (e.g., Neuron, Glia, Hep).
\item \code{"surrogate_biomarkers"}: Proxies for plasma proteins, lifestyle traits, and clonal hematopoiesis (CHIP).
}}

\item{PCClocks_RData}{A list object containing PC-Clock data (loaded via \code{load_OmniAgeR_data}).
Required if requesting PC-based clocks (e.g., PCHorvath, PCGrimAge).}

\item{SystemsAge_RData}{A list object containing SystemsAge data (loaded via \code{load_OmniAgeR_data}).
Required if requesting SystemsAge.}
}
\value{
A named list where each element is a character vector
of CpG probes for a requested clock.

A named \code{list} where each element is a \code{character vector} containing the CpG probe IDs required for a specific model.
}
\description{
Retrieves the lists of CpG probes used by the clocks specified
in the \code{clock_names} argument.

This function extracts the specific CpG probes used to construct various DNA methylation-based models.
These models include:
\itemize{
\item \strong{Epigenetic Clocks}: Chronological, biological, and mitotic age estimators (e.g., Horvath, GrimAge).
\item \strong{Surrogate Biomarkers}: DNAm predictors for plasma proteins (e.g., CRP, IL6), mitotic rates, telomere length and CHIP score.
\item \strong{Phenotypic Traits}: Estimators for lifestyle factors (e.g., BMI, Smoking).
}
Use this function to filter your methylation matrix (beta values) to only the necessary probes before calculating scores.
}
\examples{
# Get probes for all clocks
PCClocks_RData <- load_OmniAgeR_data(object_name = "PCClocks_data")
SystemsAge_RData <- load_OmniAgeR_data(object_name = "SystemsAge_data")
probes <- getClockProbes(PCClocks_RData = PCClocks_RData, SystemsAge_RData = SystemsAge_RData)
# 1. Get probes for a specific chronological aging predictor
chronological_probes <- getClockProbes(clock_names = "Horvath2013")

# 2. Get probes for a specific biological aging clocks
bio_probes <- getClockProbes(clock_names = "GrimAge1")

# 3. Get probes for all biomaker
PCClocks_RData <- load_OmniAgeR_data(object_name = "PCClocks_data")
SystemsAge_RData <- load_OmniAgeR_data(object_name = "SystemsAge_data")
all_probes <- getClockProbes(clock_names = "all", PCClocks_RData, SystemsAge_RData)

}
