% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_PCClocks_SystemsAge.R
\name{getOmniAgeRPath}
\alias{getOmniAgeRPath}
\title{Get Default Data Storage Directory}
\usage{
getOmniAgeRPath()
}
\value{
A character string representing the file path to the
package's data directory.
}
\description{
Retrieves the OS-specific, cross-platform data directory for the
'OmniAgeR' package using \code{tools::R_user_dir}.
Creates the directory if it does not already exist.
}
