% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GladyshevOrganAge.R
\name{gladyshevOrganAge}
\alias{gladyshevOrganAge}
\title{Predict Gladyshev Organ-Specific Biological Age or Mortality Risk}
\usage{
gladyshevOrganAge(
  protExp,
  modelType = c("gen2", "gen1"),
  platform = c("full", "reduced"),
  fold = 1,
  toYears = FALSE,
  standardize = TRUE
)
}
\arguments{
\item{protExp}{A numeric matrix or data frame of protein expression values.
Rows must represent protein identifiers (e.g., Gene Symbol/UniProt) and
columns must represent individual samples..}

\item{modelType}{A character string indicating the prediction objective: \code{"gen1"}
(chronological age prediction) or \code{"gen2"} (mortality risk assessment).
Default is \code{"gen2"}.}

\item{platform}{A character string denoting the proteomic assay platform used
for model training: \code{"full"} (Olink Explore 3072) or \code{"reduced"}
(Olink Explore 1536, suitable for longitudinal or external cohorts). Default is \code{"full"}.}

\item{fold}{An integer (1-5) specifying the cross-validation fold to apply,
or \code{"ensemble"} to utilize the averaged coefficients across all folds
for enhanced predictive robustness. Default is 1.}

\item{toYears}{A logical scalar. Applicable exclusively to \code{"gen2"} models.
If \code{TRUE}, converts the predicted relative log-mortality hazard scores
into biological age units (years) under a Gompertz mortality assumption. Default is \code{FALSE}.}

\item{standardize}{A logical scalar. If \code{TRUE}, performs variance alignment by
standardizing the input data and rescaling it to the standard deviations of the
UK Biobank training cohort. This step is highly recommended to mitigate batch
effects unless the data has been rigorously harmonized prior to input. Default is \code{TRUE}.}
}
\value{
A data frame containing the predicted scores for each organ.
}
\description{
This function computes organ-specific biological ages or mortality risk scores
utilizing the proteomic aging clocks developed by Goeminne et al. (2024). It maps
user-provided protein expression profiles to pre-trained Elastic Net models, offering
built-in variance alignment to the UK Biobank training cohort and Gompertz-based
transformation of mortality hazards into chronological age equivalents.
}
\examples{
protExample <- loadOmniAgeRdata(
    "omniager_proteomic_olink1536_example",
    verbose = FALSE
)

protExp <- protExample[[1]]

results <- gladyshevOrganAge(protExp,modelType="gen2",platform = "reduced")

}
\references{
Goeminne L et al.
Plasma protein-based organ-specific aging and mortality models unveil
diseases as accelerated aging of organismal systems
\emph{Cell Metabolism} 2024
}
