% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hannum.R
\name{Hannum}
\alias{Hannum}
\title{Calculate Hannum's DNAm Age (2013)}
\usage{
Hannum(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to
the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Hannum epigenetic clock age (2013) from a DNA methylation beta value matrix.
}
\details{
Implements the Hannum (2013) blood-specific clock. The function calculates
a weighted linear predictor from 71 CpGs found in the input matrix.
This clock is a direct linear model without non-linear transformation.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Hannum.out <- Hannum(hannum_bmiq_m)
}
\references{
Hannum G, Guinney J, Zhao L, et al.
Genome-wide methylation profiles reveal quantitative views of human aging rates.
\emph{Mol Cell.} 2013
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate
multiple clocks simultaneously.
}
