% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath2013.R
\name{Horvath2013}
\alias{Horvath2013}
\title{Calculate Horvath's DNAm Age (2013)}
\usage{
Horvath2013(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Horvath epigenetic clock age(2013) from a DNA methylation beta value matrix.
}
\details{
Implements the Horvath (2013) pan-tissue clock. The function calculates a weighted linear predictor from 353 CpGs found in the input matrix and then transforms this value using a non-linear function to return the final DNAm age.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Horvath2013.out <- Horvath2013(hannum_bmiq_m)

}
\references{
Horvath S.
DNA methylation age of human tissues and cell types.
\emph{Genome Biol.} 2013
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate multiple clocks simultaneously.
}
