% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath2018.R
\name{Horvath2018}
\alias{Horvath2018}
\title{Calculate Horvath's Skin & Blood DNAm Age (2018)}
\usage{
Horvath2018(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes
and columns should be individual samples.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding
to the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Horvath "Skin & Blood" clock age (2018)
from a DNA methylation beta value matrix.
}
\details{
Implements the Horvath (2018) skin & blood clock. The function calculates
a weighted linear predictor from 391 CpGs and then transforms this value
using the same non-linear function as the Horvath (2013) clock to
return the final DNAm age.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Horvath2018.out <- Horvath2018(hannum_bmiq_m)

}
\references{
Horvath S, Oshima J, Martin GM, et al.
Epigenetic clock for skin and blood cells applied to Hutchinson Gilford Progeria Syndrome and ex vivo studies.
\emph{Aging} 2018
}
