% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_Clock.R
\name{IC_Clock}
\alias{IC_Clock}
\title{Calculate the Intrinsic Capacity (IC) Epigenetic Clock}
\usage{
IC_Clock(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values.
\strong{Rows must correspond to samples and columns to CpGs.}
\code{rownames} (sample IDs) and \code{colnames} (CpG probe IDs) are required.
The matrix should not contain \code{NA} values.}
}
\value{
A \strong{numeric vector} containing the predicted Intrinsic Capacity (IC) score
for each sample. The vector is named with the sample IDs from the \code{rownames}
of \code{beta.m}.
}
\description{
Implements the DNA methylation-based predictor of Intrinsic Capacity (IC_Clock),
a biomarker of functional aging and mortality risk.
}
\details{
This function calculates the Intrinsic Capacity (IC) Clock, a novel
DNAm biomarker trained on clinical evaluations of physical and mental
capacities
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
IC_Clock.o <- IC_Clock(hannum_bmiq_m)
}
\references{
Fuentealba M, Rouch L, Guyonnet S, et al.
A blood-based epigenetic clock for intrinsic capacity predicts mortality and is associated with clinical, immunological and lifestyle factors.
\emph{Nature Aging.} 2025
}
