% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knight_GA.R
\name{Knight_GA}
\alias{Knight_GA}
\title{Calculate the Knight gestational age}
\usage{
Knight_GA(beta.m)
}
\arguments{
\item{beta.m}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}
}
\value{
A named vector of predicted Gestational Ages (in weeks).
}
\description{
A function to calculate the Knight gestational age
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Knight_GA.out <- Knight_GA(hannum_bmiq_m)
}
\references{
Knight AK, Craig JM, Theda C, et al.
An epigenetic clock for gestational age at birth based on blood methylation data.
\emph{Genome Biol.} 2016
}
