% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lee_GA.R
\name{Lee_GA}
\alias{Lee_GA}
\title{Calculate the Lee gestational age}
\usage{
Lee_GA(beta.m)
}
\arguments{
\item{beta.m}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}
}
\value{
A \code{list} containing three named numeric vectors. Each vector represents the predicted gestational age (in weeks) for the corresponding samples.
\itemize{
\item \strong{\code{LeeControl}}: Numeric vector of predicted GAs from the Control model.
\item \strong{\code{LeeRobust}}: Numeric vector of predicted GAs from the Robust model.
\item \strong{\code{LeeRefinedRobust}}: Numeric vector of predicted GAs from the Refined Robust model.
}
Each vector is named with the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Implements the placental epigenetic clocks for estimating gestational age (GA) using DNA methylation data, as described by Lee et al. (2019).
}
\details{
This function computes three distinct GA clocks derived from the models presented in the Lee et al. (2019) study.

The implemented clocks are:
\itemize{
\item \strong{\code{LeeControl}}: The control model (546 CpGs).
\item \strong{\code{LeeRobust}}: The robust model (558 CpGs).
\item \strong{\code{LeeRefinedRobust}}: The refined robust model (395 CpGs).
}
The function iterates through each clock, matches the required CpGs (e.g., 546 for \code{LeeControl}) with the columns in the input \code{beta.m} matrix, and calculates a linear prediction of GA.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Lee_GA.out <- Lee_GA(hannum_bmiq_m)
}
\references{
Lee Y, Choufani S, Weksberg R, et al.
Placental epigenetic clocks: estimating gestational age using placental DNA methylation levels.
\emph{Aging} 2019
}
