% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin.R
\name{Lin}
\alias{Lin}
\title{Calculate Lin DNAm Age (2016)}
\usage{
Lin(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to
the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Lin epigenetic clock age (2016) from a DNA methylation beta value matrix.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Lin.out <- Lin(beta.m = hannum_bmiq_m)

}
\references{
Lin Q, Weidner CI, Costa IG, et al.
DNA methylation levels at individual age-associated CpG sites can be indicative for life expectancy.
\emph{Aging} 2016
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate
multiple clocks simultaneously.
}
