% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalAgingScore.R
\name{listEpiGA}
\alias{listEpiGA}
\title{List available Gestational Age (GA) epigenetic clocks}
\usage{
listEpiGA()
}
\value{
A character vector where each element is a string corresponding
to the name of an implemented GA clock.
}
\description{
Retrieves a character vector containing the names of all implemented
Gestational Age (GA) epigenetic clocks available within the package.
This utility function is primarily used to identify the valid clock names
(strings) that can be supplied to the \code{clock_names} argument in other
functions, such as \code{EpiGA}.
}
\examples{
available_ga_clocks <- listEpiGA()
}
