% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_OmniAgeR_example}
\alias{load_OmniAgeR_example}
\title{Load OmniAgeR Example Data}
\usage{
load_OmniAgeR_example(dataset_name, envir = parent.frame())
}
\arguments{
\item{dataset_name}{Base name of the dataset (e.g., "seu_gabitto_2024_filtered").}

\item{envir}{The environment where the data should be loaded. Defaults to parent.frame().}
}
\description{
Safely load .rda example datasets from the local storage path.
}
