% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadOmniAgeRExample}
\alias{loadOmniAgeRExample}
\title{Load OmniAgeR Example Data}
\usage{
loadOmniAgeRExample(datasetName, envir = parent.frame())
}
\arguments{
\item{envir}{The environment where the data should be loaded. Defaults to parent.frame().}

\item{dataset_name}{Base name of the dataset (e.g., "seu_gabitto_2024_filtered").}
}
\value{
Invisible NULL. The dataset is loaded into the specified environment as a side effect.
}
\description{
Safely load .rda example datasets from the local storage path.
}
\examples{
loadOmniAgeRExample("Hannum_example")
}
