% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mayne_GA.R
\name{Mayne_GA}
\alias{Mayne_GA}
\title{Calculate the Mayne Placental Gestational Age.}
\usage{
Mayne_GA(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values.
\strong{Rows must correspond to samples and columns to CpGs.}
\code{rownames} (sample IDs) and \code{colnames} (CpG probe IDs) are required.
The matrix should not contain \code{NA} values.}
}
\value{
A \strong{numeric vector} containing the predicted gestational age (in weeks) for each sample. The vector is named with the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Implements the 62-CpG placental epigenetic clock for estimating gestational age (GA), as described by Mayne et al. (2017).
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Mayne_GA.out <- Mayne_GA(hannum_bmiq_m)
}
\references{
Mayne BT, Leemaqz SY, Smith AK, Breen J, Roberts CT, Bianco-Miotto T.
Accelerated placental aging in early onset preeclampsia pregnancies identified by DNA methylation.
\emph{Epigenomics} 2017
}
