% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmniAgeR-package.R
\docType{package}
\name{OmniAgeR-package}
\alias{OmniAgeR-package}
\alias{OmniAgeR}
\title{OmniAgeR: A Comprehensive Suite for Aging Omic Biomarker}
\description{
The \code{OmniAgeR} package provides a standardized, efficient, and comprehensive
suite of tools for calculating and evaluating various aging omic biomarkers
}
\details{
\code{OmniAgeR} simplifies the application of complex aging omic biomarkers by
providing a unified interface (e.g., supporting matrices, \code{SummarizedExperiment},
and \code{SingleCellExperiment} objects) and automating data preprocessing.

For detailed tutorials and pipeline workflows, please refer to the package vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Duzhaozhen/OmniAgeR}
  \item Report bugs at \url{https://github.com/Duzhaozhen/OmniAgeR/issues}
}

}
\author{
\strong{Maintainer}: Zhaozhen Du \email{duzhaozhen2022@sinh.ac.cn} (affiliation: Shanghai Institute of Nutrition and Health, CAS)

Authors:
\itemize{
  \item Zhaozhen Du \email{duzhaozhen2022@sinh.ac.cn} (affiliation: Shanghai Institute of Nutrition and Health, CAS)
  \item Andrew E. Teschendorff \email{andrew@sinh.ac.cn} (affiliation: Shanghai Institute of Nutrition and Health, CAS)
}

Other contributors:
\itemize{
  \item National Science Foundation of China (NSFC: W2431024, 32370699, 32570775) [funder]
}

}
\keyword{internal}
