% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASTA.R
\name{PASTA_Scores}
\alias{PASTA_Scores}
\title{Calculate PASTA-based transcriptomic age scores}
\usage{
PASTA_Scores(
  mat,
  filter_genes = TRUE,
  rank_norm = TRUE,
  REG = TRUE,
  PASTA = TRUE,
  CT46 = TRUE
)
}
\arguments{
\item{mat}{An expression matrix (numeric) with \strong{genes as rows} and
\strong{samples as columns}.}

\item{filter_genes}{Logical. If \code{TRUE} (default), the matrix is
subsetted to retain only the genes utilized by the pre-trained models.}

\item{rank_norm}{Logical. If \code{TRUE} (default), applies a rank-based
inverse normal transformation (rank-normalization) to the expression data.}

\item{REG}{Logical. If \code{TRUE} (default), computes the REG (regression)
age score.}

\item{PASTA}{Logical. If \code{TRUE} (default), computes the PASTA age score.}

\item{CT46}{Logical. If \code{TRUE} (default), computes the CT46 age score.}
}
\value{
A \code{list} where each element is a numeric vector of predicted age scores
for the requested model(s) (e.g., \code{res_list$PASTA}, \code{res_list$REG}).
}
\description{
This function computes transcriptomic age acceleration scores based on the
models described by Salignon et al. (2025), including the primary PASTA
score, a standard regression (REG) score, and the CT46 score.
}
\examples{
library(magrittr)
library(Seurat)
library(glmnet)
download_OmniAgeR_example("seu_gabitto_2024_filtered")
load_OmniAgeR_example("seu_gabitto_2024_filtered")

seu$age <- seu$development_stage \%>\%
 gsub("-year.*", "", .) \%>\%
 gsub("-", " ", .) \%>\%
 gsub("80 year old and over stage", "85", .)

set.seed(42)
seu_bulk <- making_pseudobulks_from_seurat(seu,pool_by = c("cell_type", "age"), chunk_size = 512, verbose = FALSE)

lognorm_matrix <- GetAssayData(seu_bulk, assay = "RNA", layer = "data")
lognorm_matrix <- as.matrix(lognorm_matrix)

seu_bulk_meta <- seu_bulk[[c('chunk_size', 'cell_type', 'age')]]
seu_bulk_meta$age <- as.numeric(seu_bulk_meta$age)
PASTA_res <- PASTA_Scores(lognorm_matrix, filter_genes = TRUE, rank_norm = TRUE)

}
\references{
Salignon, J., Tsiokou, M., Marqués, P. et al.
Pasta, an age-shift transcriptomic clock, maps the chemical and genetic determinants of aging and rejuvenation.
\emph{bioRxiv.} 2025
}
