% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedBE.R
\name{PedBE}
\alias{PedBE}
\title{The PedBE (Pediatric Buccal) Clock for DNAm Age in Children}
\usage{
PedBE(beta.m)
}
\description{
Implements the Pediatric Buccal Epigenetic (PedBE) clock, specifically developed to estimate DNA methylation age in \strong{pediatric (childhood) samples}, as described by McEwen et al. (2020).
}
\details{
This clock is specifically trained on and designed for pediatric buccal
epithelial (cheek swab) samples. The calculation is a two-step process:
\enumerate{
\item A linear predictor is first calculated from the beta values using the 94-CpG elastic net coefficients . This value represents a \emph{transformed} age.
\item A non-linear inverse age transformation (via the internal \code{anti.trafo} function from Horvath, 2013) is then applied. This converts the transformed age into a final estimate of chronological age in years.
}
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
PedBE.out <- PedBE(hannum_bmiq_m)
}
\references{
McEwen LM, O'Donnell KJ, McGill MG, et al.
The PedBE clock accurately estimates DNA methylation age in pediatric buccal cells.
\emph{Proc Natl Acad Sci U S A.} 2020
}
