% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PetkovichMouse.R
\name{petkovichMouse}
\alias{petkovichMouse}
\title{Petkovich Blood Mouse DNA Methylation Age Predictor}
\usage{
petkovichMouse(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predictions.
}
\description{
Predicts the biological age of mouse based on the blood-based epigenetic clock
developed by Petkovich et al. (2017). This model utilizes the methylation
status of 90 specific CpG sites to estimate age.
}
\details{
This predictor was constructed using \strong{Elastic Net regression} on partial
DNA methylomes obtained via Reduced Representation Bisulfite Sequencing (RRBS)
from a cohort of 141 C57BL/6 mice.

The estimation process involves:
\enumerate{
\item \strong{Linear Summation}: Calculating a raw DNAm score (MSc) by taking the
weighted average of methylation levels across 90 CpG sites.
\item \strong{Non-linear Transformation}: Applying a \strong{Power Law} conversion (i.e.
((MSc - c) / a)^(1/b) ) to account for the decelerating rate of
epigenetic changes observed during  mouse maturation.
\item \strong{Output}: Inverting the power law function to derive the predicted
"Methylation Age".
}
}
\examples{
exampleMat <- loadOmniAgeRdata(
    "omniager_mouse_rrbs_example",
    verbose = FALSE
)
predRes <- petkovichMouse(exampleMat)

}
\references{
Petkovich, Daniel A et al.
Using DNA Methylation Profiling to Evaluate Biological Age and Longevity Interventions.
\emph{Cell Metab.} 2017
}
