% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASTA.R
\name{predictAgeScore}
\alias{predictAgeScore}
\title{Predict Age Score from Gene Expression Matrix}
\usage{
predictAgeScore(mat, modelType = "PASTA")
}
\arguments{
\item{mat}{Matrix. Processed count matrix.}

\item{modelType}{Character. Model type ('PASTA', 'REG', or 'CT46').}
}
\value{
Numeric vector. Predicted age scores.
}
\description{
Uses pre-trained models to predict age scores based on gene expression.
}
