% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASTA.R
\name{predicting_age_score}
\alias{predicting_age_score}
\title{Predict Age Score from Gene Expression Matrix}
\usage{
predicting_age_score(mat, model_type = "PASTA")
}
\arguments{
\item{mat}{Matrix. Processed count matrix.}

\item{model_type}{Character. Model type ('PASTA', 'REG', or 'CT46').}
}
\value{
Numeric vector. Predicted age scores.
}
\description{
Uses pre-trained models to predict age scores based on gene expression.
}
