% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuckleyMouseSVZ.R
\name{predictSCRNAClock}
\alias{predictSCRNAClock}
\title{General Clock Score Calculator (Supports Dynamic Metadata & Imputation)}
\usage{
predictSCRNAClock(
  preprocessedData,
  weightDf,
  minCoverage = 0,
  imputeData = NULL,
  clockName,
  metaColNames = c("donorId", "age"),
  sampleType = "Bootstrap",
  verbose = TRUE
)
}
\arguments{
\item{preprocessedData}{A data frame/tibble output from pre-processing.}

\item{weightDf}{A data frame of weights with 'feature_name' and 'coefficient'.}

\item{minCoverage}{Minimum proportion of required genes (default 0).}

\item{imputeData}{Optional. A data frame with 'feature_name' and 'imputation_value'
to fill in missing genes. If NULL, missing genes are filled with 0.}

\item{clockName}{Character for logging.}

\item{metaColNames}{Character vector specifying which columns are metadata.
Default is c("donorId", "age").}

\item{sampleType}{A character string (e.g., 'SC', 'Pseudobulk') used to
tag the output data.frame.}

\item{verbose}{Logical flag.}
}
\value{
Return dataframe of predicted scores safely merged with metadata.
}
\description{
General Clock Score Calculator (Supports Dynamic Metadata & Imputation)
}
