% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retro_age.R
\name{Retro_age}
\alias{Retro_age}
\title{Calculate the Retro-age Epigenetic Clock}
\usage{
Retro_age(beta.m, version = "V2", verbose = TRUE)
}
\arguments{
\item{beta.m}{A matrix of DNA methylation beta values. \strong{Rows must correspond
to samples} and \strong{columns to CpG probes}. Rownames (sample IDs) and
colnames (CpG probe IDs) are required.}

\item{version}{A character string specifying which version of the clock to
use. Valid options are "V1" (EPICv1) or "V2" (EPICv1 & EPICv2 compatible,
recommended). Defaults to "V2".}

\item{verbose}{A logical value. If TRUE (default), the function will
print messages detailing the calculation steps.}
}
\value{
A numeric vector of predicted ages (Retro-age). The vector is
named with the sample IDs from the input matrix rownames.
}
\description{
Calculates the "Retro-age," a retroelement-based epigenetic clock for
chronological age, based on the models developed by Ndhlovu et al. (2024).
The function can compute either Version 1 (V1) or Version 2 (V2) of the clock.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Retro_age.o <- Retro_age(hannum_bmiq_m)
}
\references{
Ndhlovu LC, Bendall ML, Dwaraka V, et al.
Retro-age: A unique epigenetic biomarker of aging captured by DNA methylation states of retroelements.
\emph{Aging Cell.} 2024
}
