% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retro_age.R
\name{retroAge}
\alias{retroAge}
\title{Calculate the Retro-age Epigenetic Clock}
\usage{
retroAge(x, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{x}{A numeric matrix, \code{data.frame}, or \code{SummarizedExperiment}
object containing DNA methylation beta values. Rows should be CpG probes and
columns individual samples.}

\item{minCoverage}{A numeric value between 0 and 1 (default is 0).
Specifies the minimum proportion of required CpGs that must be present
in the input matrix for the clock calculation to proceed.}

\item{verbose}{A logical value. If TRUE (default), the function will
print messages detailing the calculation steps.}
}
\value{
A list containing the predicted age for the "V1" and "V2" clocks.
}
\description{
Calculates the "Retro-age," a retroelement-based epigenetic clock for
chronological age, based on the models developed by Ndhlovu et al. (2024).
This function computes both Version 1 (V1) and Version 2 (V2) of the clock.
}
\examples{
data(dnamExample)
beta_matrix <- dnamExample[[1]]

# Example 1: Direct Matrix Input
predOut <- retroAge(x = beta_matrix, verbose = FALSE)

# Example 2: SummarizedExperiment Input
\dontrun{
  library(SummarizedExperiment)
  pheno_data <- dnamExample[[2]]
  rownames(pheno_data) <- colnames(beta_matrix)

  se_obj <- SummarizedExperiment(
    assays = list(beta = beta_matrix),
    colData = pheno_data
  )

  predOut <- retroAge(x = se_obj, verbose = FALSE)
}

}
\references{
Ndhlovu LC, Bendall ML, Dwaraka V, et al.
Retro-age: A unique epigenetic biomarker of aging captured by DNA methylation states of retroelements.
\emph{Aging Cell.} 2024
}
