% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_scImmuAging.R
\name{scImmuAging_PreProcess}
\alias{scImmuAging_PreProcess}
\title{Pre-Processing of input dataset}
\usage{
scImmuAging_PreProcess(input, cell_type, model, marker_gene)
}
\arguments{
\item{input}{A Seurat object. Its metadata (meta.data) must contain "donor_id" and "age" columns.}

\item{cell_type}{A string specifying the cell type to be analyzed.}

\item{model}{A string or object specifying the model to be used.}

\item{marker_gene}{selected marker genes from the corresponding model.}
}
\value{
Return pre-processed input
}
\description{
input is a seurat object. meta.data should have "donor_id" and "age" columns.
}
