% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_scImmuAging.R
\name{scImmuAgingPreprocessing}
\alias{scImmuAgingPreprocessing}
\title{Pre-Processing of input dataset}
\usage{
scImmuAgingPreprocessing(seuratObj)
}
\arguments{
\item{seuratObj}{seurat object with metadata "donor_id", "age"}
}
\value{
Return pre-processed input
}
\description{
input is a seurat object. meta.data should have "donor_id" and "age" columns.
}
