% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemTOC.R
\name{stemTOC}
\alias{stemTOC}
\title{Estimate stemTOC score}
\usage{
stemTOC(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows). Rows
should be Illumina 450k/EPIC CpG identifiers and columns should be samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
The stemTOC score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and
will return the stemTOC score.
}
\details{
The function will return the 0.95 upper quantile of the 371 stemTOC CpGs.
Compared to stemTOCvitro CpGs, the stemTOC CpGs are filtered for
significant DNA hypermethylation with chronological age
in large in-vivo datasets
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
stemtocOut <- stemTOC(betaM = lungInvM)
}
\references{
Zhu, T., Tong, H., Du, Z. et al.
An improved epigenetic counter to track mitotic age in normal
and precancerous tissues.
\emph{Nat Commun} 2024
}
