% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bohlin_GA.R
\name{bohlinGa}
\alias{bohlinGa}
\title{Calculate the Bohlin Gestational Age (Cord Blood)}
\usage{
bohlinGa(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{a matrix of methylation beta values.
Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
#' A \strong{numeric vector} containing the predicted gestational
age (in weeks) for each sample. The vector is named with the sample IDs
from the \code{rownames} of \code{betaM}.
}
\description{
Implements the epigenetic clock for predicting gestational age (GA) using
newborn cord blood, as described by Bohlin et al. (2016).
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
bohlinGaOut <- bohlinGa(hannumBmiqM)
}
\references{
Bohlin J, Håberg SE, Magnus P, et al.
Prediction of gestational age based on genome-wide differentially
methylated regions. \emph{Genome Biol.} 2016
}
