% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalClocks.R
\name{CausalClock}
\alias{CausalClock}
\title{Calculate Causal, Damage, and Adaptation Epigenetic Clocks}
\usage{
CausalClock(beta.m)
}
\arguments{
\item{beta.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
A list containing the predicted scores for the "Causal", "Damage", and "Adaptation" clocks.
}
\description{
Calculates three related epigenetic clocks (Causal, Damage, Adaptation) from a DNA methylation beta value matrix. These clocks were developed to distinguish different aspects of aging.
}
\details{
This function implements the three epigenetic clocks described by Ying et al. (2024) to dissect biological aging into distinct components.
The Damage clock (DamAge) is designed to capture the accumulation of molecular damage.
The Adaptation clock (AdaptAge) reflects the body's adaptive responses to this damage.
The Causal clock (CausalAge) is enriched for CpGs with a causal effect on mortality.
Each score is calculated as a weighted linear sum of beta values from its specific set of CpG sites.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
CausalClock.out <- CausalClock(hannum_bmiq_m)

}
\references{
Ying K, Liu H, Tarkhov AE, et al.
Causality-enriched epigenetic age uncouples damage and adaptation.
\emph{Nat Aging} 2024
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate multiple clocks simultaneously.
}
