% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentenarianClock.R
\name{CentenarianClock}
\alias{CentenarianClock}
\title{Calculate Centenarian Epigenetic Clocks (Eric Dec et al.)}
\usage{
CentenarianClock(beta.m, verbose = TRUE)
}
\arguments{
\item{beta.m}{a matrix of methylation beta values. Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{verbose}{A boolean (default: \code{TRUE}). If \code{TRUE}, the function will
print status messages, including the number of represented CpGs
for each clock.}
}
\value{
A list containing the predicted scores for each Centenarian clock.
\itemize{
\item \code{ENCen40}: Elastic net clock trained on individuals aged 40+.
\item \code{ENCen100}: Elastic net clock specifically trained on centenarians (100+).
}
}
\description{
Calculates the Centenarian epigenetic clocks (ENCen40 and ENCen100) developed by Eric Dec et al. (2023). This function serves as a wrapper that loads the internal clock coefficients and computes the linear predictors for each clock using the helper function.
\code{calculateLinearPredictor}.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
CentenarianClock.out <- CentenarianClock(hannum_bmiq_m)
}
\references{
Dec, E., Clement, J., Cheng, K. et al.
Centenarian clocks: epigenetic clocks for validating claims of exceptional longevity.
\emph{GeroScience} 2023
}
