% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentenarianClock.R
\name{centenarianClock}
\alias{centenarianClock}
\title{Calculate Centenarian Epigenetic Clocks (Eric Dec et al.)}
\usage{
centenarianClock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{a matrix of methylation beta values.
Needs to be rows = samples and columns = CpGs, with rownames and colnames.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A list containing the predicted scores for each Centenarian clock.
\itemize{
\item \code{ENCen40}: Elastic net clock trained on individuals aged 40+.
\item \code{ENCen100}: Elastic net clock specifically trained on centenarians (100+).
}
}
\description{
Calculates the Centenarian epigenetic clocks
(ENCen40 and ENCen100) developed by Eric Dec et al. (2023). This function
serves as a wrapper that loads the internal clock coefficients and computes
the linear predictors for each clock using the helper function.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
CentenarianClock.out <- centenarianClock(hannum_bmiq_m)

}
\references{
Dec, E., Clement, J., Cheng, K. et al.
Centenarian clocks: epigenetic clocks for validating claims of
exceptional longevity. \emph{GeroScience} 2023
}
