% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCHIP.R
\name{compCHIP}
\alias{compCHIP}
\title{Calculate CHIP-related Methylation Scores}
\usage{
compCHIP(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \code{list} of numeric vectors, one for each CHIP signature.
}
\description{
Computes CHIP (Clonal Hematopoiesis of Indeterminate Potential) scores
based on predefined CpG signatures and a normalized DNA methylation matrix.
}
\details{
For each CHIP signature (e.g., DNMT3A, TET2), it identifies common CpGs,
calculates Z-scores for each
probe across all samples, and then computes the score as:
(Mean Z-score of positive probes) - (Mean Z-score of negative probes).

Probes with zero variance across samples are excluded.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
compchipOut <- compCHIP(hannumBmiqM)
}
\references{
Kirmani, S., Huan, T., Van Amburg, J.C. et al.
Epigenome-wide DNA methylation association study of CHIP provides insight
into perturbed gene regulation.
\emph{Nat Commun} 2025
}
