% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCHIP.R
\name{compCHIP}
\alias{compCHIP}
\title{Calculate CHIP-related Methylation Scores}
\usage{
compCHIP(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows). Rows
should be Illumina 450k/EPIC CpG identifiers and columns should be samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.5.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
A \code{list} of numeric vectors, one for each CHIP signature.
}
\description{
Computes CHIP (Clonal Hematopoiesis of Indeterminate Potential) scores
based on predefined CpG signatures and a normalized DNA methylation matrix.
}
\details{
For each CHIP signature (e.g., DNMT3A, TET2), it identifies common CpGs,
calculates Z-scores for each
probe across all samples, and then computes the score as:
(Mean Z-score of positive probes) - (Mean Z-score of negative probes).

Probes with zero variance across samples are excluded.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
compchipOut <- compCHIP(hannumBmiqM)
}
\references{
Kirmani, S., Huan, T., Van Amburg, J.C. et al.
Epigenome-wide DNA methylation association study of CHIP provides insight
into perturbed gene regulation.
\emph{Nat Commun} 2025
}
