% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCRP.R
\name{CompCRP}
\alias{CompCRP}
\title{Compute a DNAm-based C-Reactive Protein (CRP) Score}
\usage{
CompCRP(nbeta.m)
}
\arguments{
\item{nbeta.m}{normalized beta-valued DNAm data matrix with rownames the CpG identifiers. Missing values should be imputed.}
}
\value{
A named list containing two numeric vectors of CRP estimates, one for each signature provided in \code{crpCpG.lv} (\code{crp} and \code{intCRP}).
}
\description{
Computes a DNAm-based surrogate proxy for C-Reactive Protein (CRP) levels, a key biomarker for systemic inflammation (inflammaging).
}
\details{
This function calculates a DNAm-based CRP proxy score from a normalized beta-valued DNAm data matrix. The score is derived by correlating the methylation values with the signs of pre-defined CpG weights from two internal signatures: (i) crp: a vector containing 1765 CpGs and weights associated with CRP as inferred from a large meta-analysis over 20,000 blood samples where both DNAm and CRP was measured, and where no adjustment for variations between memory and naive T-cells was made. (ii) intCRP: contains a 62 CpG CRP-signature that is adjusted for variations between memory and naive T-cells. The resulting output is a relative score that captures the methylation pattern associated with CRP levels.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
CompCRP.out <- CompCRP(hannum_bmiq_m)
}
\author{
Andrew Teschendorff (andrew@sinh.ac.cn)
}
