% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCRP.R
\name{compCRP}
\alias{compCRP}
\title{Compute a DNAm-based C-Reactive Protein (CRP) Score}
\usage{
compCRP(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows). Rows
should be Illumina 450k/EPIC CpG identifiers and columns should be samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
A \code{list} containing two numeric vectors:
\code{CRP} and \code{intCRP}.
}
\description{
Computes a DNAm-based surrogate proxy for C-Reactive Protein (CRP) levels,
a key biomarker for systemic inflammation (inflammaging).
}
\details{
This function calculates a DNAm-based CRP proxy score from a normalized
beta-valued DNAm data matrix. The score is derived by correlating the
methylation values with the signs of pre-defined CpG weights from two
internal signatures: (i) crp: a vector containing 1765 CpGs and weights
associated with CRP as inferred from a large meta-analysis over 20,000 blood
samples where both DNAm and CRP was measured, and where no adjustment for
variations between memory and naive T-cells was made. (ii) intCRP: contains
a 62 CpG CRP-signature that is adjusted for variations between memory and
naive T-cells. The resulting output is a relative score that captures the
methylation pattern associated with CRP levels.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
compcrpOut <- compCRP(hannumBmiqM)
}
\references{
Wielscher, M., Mandaviya, P.R., Kuehnel, B. et al.
DNA methylation signature of chronic low-grade inflammation and its role in
cardio-respiratory diseases.
\emph{Nat Commun} 2022
}
