% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompIL6.R
\name{CompIL6}
\alias{CompIL6}
\title{Calculate a DNA Methylation-Based Proxy for IL-6}
\usage{
CompIL6(beta.m)
}
\arguments{
\item{nbeta.m}{normalized beta-valued DNAm data matrix with rownames the CpG identifiers. Missing values should be imputed.}
}
\value{
A numeric vector containing the calculated IL-6 proxy score for each sample. The vector is named according to the column names (sample IDs) of the input \code{beta.m} matrix.
}
\description{
Computes a DNA methylation (DNAm) surrogate score for Interleukin-6 (IL-6) protein levels. IL-6 is a key cytokine and biomarker for systemic inflammation.
}
\details{
This function calculates the IL-6 proxy score by applying a pre-defined
set of coefficients to the input beta-value matrix. The coefficients
are loaded from the internal \code{IL6Coef} data object.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
CompIL6.out <- CompIL6(hannum_bmiq_m)
}
