% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompIL6.R
\name{compIL6}
\alias{compIL6}
\title{Calculate a DNA Methylation-Based Proxy for IL-6}
\usage{
compIL6(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector containing the calculated IL-6 proxy score for each
sample. The vector is named according to the column names (sample IDs) of
the input matrix.
}
\description{
Computes a DNAm surrogate score for Interleukin-6 (IL-6)
protein levels.
}
\details{
This function calculates the IL-6 proxy score by applying a pre-defined
set of coefficients to the input beta-value matrix. The coefficients
are loaded from the internal \code{IL6Coef} data object.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
compil6Out <- compIL6(hannum_bmiq_m)
}
\references{
Stevenson AJ et al.
Creating and Validating a DNA Methylation-Based Proxy for Interleukin-6.
\emph{J Gerontol A Biol Sci Med Sci.} 2021
}
