% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSmokeIndex.R
\name{compSmokeIndex}
\alias{compSmokeIndex}
\title{Compute a DNAm-based Smoking Index}
\usage{
compSmokeIndex(betaMat, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}

\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}
}
\value{
A named \code{numeric} vector of smoking index for each sample
in \code{data.m}.
}
\description{
Computes a DNA methylation-based smoking index (S-index) using 1,501 smoking-associated
CpG sites. This index has been shown to correlate with cancer risk across multiple
tissues and provides a robust epigenetic signature of smoking exposure.
}
\details{
The function calculates the score by first standardizing the methylation data (z-score
transformation) across samples for each CpG site, and then computing a weighted
average based on the provided signature coefficients.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
CompSmokeIndex_o <- compSmokeIndex(hannum_bmiq_m)

}
\references{
Teschendorff, Andrew E et al.
Correlation of Smoking-Associated DNA Methylation Changes in Buccal Cells
With DNA Methylation Changes in Epithelial Cancer
\emph{JAMA oncology} 2015
}
