% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorticalClock.R
\name{CorticalClock}
\alias{CorticalClock}
\title{Predict Cortical DNA Methylation Clock Age (2020)}
\usage{
CorticalClock(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A named numeric vector of predicted cortical DNAm ages for each sample.
}
\description{
Predicts DNAm age in cortical samples using the elastic net model by Shireby et al. (2020).
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
CorticalClock.out <- CorticalClock(hannum_bmiq_m)

}
\references{
Shireby GL, Davies JP, Francis PT, et al.
Recalibrating the epigenetic clock: implications for assessing biological age in the human cortex.
\emph{Brain.} 2020
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate multiple clocks simultaneously.
}
