% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorticalClock.R
\name{corticalClock}
\alias{corticalClock}
\title{Predict Cortical DNA Methylation Clock Age (2020)}
\usage{
corticalClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A named numeric vector of predicted cortical DNAm ages for
each sample.
}
\description{
Predicts DNAm age in cortical samples using the elastic net
model by Shireby et al. (2020).
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
corticalClockOut <- corticalClock(hannumBmiqM)
}
\references{
Shireby GL, Davies JP, Francis PT, et al.
Recalibrating the epigenetic clock: implications for assessing
biological age in the human cortex.
\emph{Brain.} 2020
}
