% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_CTF_Clock.R
\name{dnamCTFClock}
\alias{dnamCTFClock}
\title{DNA Methylation Cell-Type Fraction (CTF) Aging Clock}
\usage{
dnamCTFClock(ctfM, verbose = TRUE)
}
\arguments{
\item{ctfM}{A numeric matrix or data frame where rows are samples and columns are cell types.
Must contain the specific cell types required by the model (e.g., predicted by EpiDISH).}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A named numeric vector of predicted ages.
}
\description{
Predicts biological age based on immune cell type fractions derived from
DNA methylation data.
}
\examples{

tzhFracM <- loadOmniAgeRdata(
    "omniager_tzh_example_ctf",
    verbose = FALSE
)[[2]]
dnamCTFClockOut <- dnamCTFClock(ctfM = tzhFracM)

}
