% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_CTF_Clock.R
\name{dnamCTFClock}
\alias{dnamCTFClock}
\title{DNA Methylation Cell-Type Fraction (CTF) Aging Clock}
\usage{
dnamCTFClock(object)
}
\arguments{
\item{ctfM}{A numeric matrix or data frame where rows are samples and columns are cell types.
Must contain the specific cell types required by the model (e.g., predicted by EpiDISH).}
}
\value{
A named numeric vector of predicted ages.
}
\description{
Predicts biological age based on immune cell type fractions derived from
DNA methylation data.
}
\examples{

downloadOmniAgeRExample("TZH_example_CTF")
loadOmniAgeRExample("TZH_example_CTF")
dnamCTFClockOut <- dnamCTFClock(ctfM = TZH_Frac_m)

}
