% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmFitAge.R
\name{DNAmFitAge}
\alias{DNAmFitAge}
\title{Calculate DNAmFitAge and related fitness biomarkers}
\usage{
DNAmFitAge(beta_matrix, age_vector, sex_vector, grimage_vector, verbose = TRUE)
}
\arguments{
\item{beta_matrix}{A numeric matrix. \strong{Rows must be CpGs, Columns must be Samples.}
The \code{colnames} must be the sample IDs.}

\item{age_vector}{A numeric vector of chronological age for each sample.
The order \strong{must match the column order} of \code{beta_matrix}.}

\item{sex_vector}{A character or factor vector of sex for each sample
(e.g., "Male" or "Female"). The order \strong{must match the column order}
of \code{beta_matrix}.}

\item{grimage_vector}{A numeric vector of pre-calculated DNAmGrimAge values.
The order \strong{must match the column order} of \code{beta_matrix}.}

\item{verbose}{Logical. If \code{TRUE} (default), the function will print
progress messages and QC information to the console.}

\item{package_name}{A string with the name of your R package
(e.g., "OmniAgeR"). This is used to locate the internal model data file.}
}
\value{
A \code{data.frame} with 12 columns:
\itemize{
\item \code{SampleID}: The sample identifiers.
\item \code{Age}: The input chronological age.
\item \code{Female}: The input sex, coded as 1 for Female, 0 for Male.
\item \code{DNAmGait_noAge}, \code{DNAmGrip_noAge}, \code{DNAmVO2max}: Fitness biomarkers.
\item \code{DNAmGait_wAge}, \code{DNAmGrip_wAge}, \code{DNAmFEV1_wAge}: Age-adjusted fitness biomarkers.
\item \code{DNAmGrimAge}: The input DNAmGrimAge.
\item \code{DNAmFitAge}: The calculated biological fitness age.
\item \code{FitAgeAccel}: The fitness age acceleration (residual of DNAmFitAge regressed on Age).
}
}
\description{
Calculates all 6 DNAm fitness biomarkers, DNAmFitAge, and FitAgeAcceleration
from a DNA methylation matrix and phenotype data.

This function serves as a wrapper that:
\enumerate{
\item Pre-processes the data and imputes missing CpGs.
\item Calls \code{internal_DNAmFitnessEstimators} to calculate 6 fitness markers.
\item Calls \code{internal_FitAgeEstimator} to calculate DNAmFitAge and its acceleration.
}
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
age <- PhenoTypesHannum_lv$Age
sex <- ifelse(PhenoTypesHannum_lv$Sex=="F","Female","Male")
GrimAge1.o <- GrimAge1(hannum_bmiq_m,age,sex)

DNAmFitAge.o <- DNAmFitAge(hannum_bmiq_m,age,sex,GrimAge1.o$DNAmGrimAge1 )
}
\references{
McGreevy KM, Radak Z, Torma F, et al.
DNAmFitAge: biological age indicator incorporating physical fitness.
\emph{Aging} 2023
}
