% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmFitAge.R
\name{dnamFitAge}
\alias{dnamFitAge}
\title{Calculate DNAmFitAge and related fitness biomarkers}
\usage{
dnamFitAge(betaM, age, sex, grimageVector, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix. \strong{Rows must be CpGs, Columns must be Samples.}
The \code{colnames} must be the sample IDs.}

\item{age}{A numeric vector of chronological age for each sample.
The order \strong{must match the column order} of \code{betaM}.}

\item{sex}{A character or factor vector of sex for each sample
(e.g., "Male" or "Female"). The order \strong{must match the column order}
of \code{beta_matrix}.}

\item{grimageVector}{A numeric vector of pre-calculated DNAmGrimAge values.
The order \strong{must match the column order} of \code{betaM}.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages}
}
\value{
A \code{data.frame} with 12 columns:
\itemize{
\item \code{SampleID}: The sample identifiers.
\item \code{DNAmGait_noAge}, \code{DNAmGrip_noAge}, \code{DNAmVO2max}: Fitness biomarkers.
\item \code{DNAmGait_wAge}, \code{DNAmGrip_wAge}, \code{DNAmFEV1_wAge}: Age-adjusted fitness biomarkers.
\item \code{DNAmGrimAge}: The input DNAmGrimAge.
\item \code{DNAmFitAge}: The calculated biological fitness age.
}
}
\description{
Calculates all 6 DNAm fitness biomarkers, DNAmFitAge, and FitAgeAcceleration
from a DNA methylation matrix and phenotype data.
}
\examples{
hannumExample <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)
hannumBmiqM <- hannumExample[[1]]
phenoTypesHannum <- hannumExample[[2]]
age <- phenoTypesHannum$Age
sex <- ifelse(phenoTypesHannum$Sex == "F", "Female", "Male")
GrimAge1O <- grimAge1(hannumBmiqM, age, sex)
dnamFitAgeOut <- dnamFitAge(hannumBmiqM, age, sex, GrimAge1O$DNAmGrimAge1)
}
\references{
McGreevy KM, Radak Z, Torma F, et al.
DNAmFitAge: biological age indicator incorporating physical fitness.
\emph{Aging} 2023
}
