% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANmTL.R
\name{DNAmTL}
\alias{DNAmTL}
\title{The epigenetic age used for calculating the Leukocyte telomere length (2019).}
\usage{
DNAmTL(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A named vector of predicted Leukocyte telomere length.
}
\description{
A function to calculate the the Leukocyte telomere length (2019) from a DNA methylation beta value matrix.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
DNAmTL.o <- DNAmTL(hannum_bmiq_m)
}
\references{
Lu AT, Seeboth A, Tsai PC, et al.
DNA methylation-based estimator of telomere length
\emph{Aging} 2019
}
