% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANmTL.R
\name{dnamTL}
\alias{dnamTL}
\title{The epigenetic age used for calculating the Leukocyte telomere length.}
\usage{
dnamTL(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}

\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}
}
\value{
A named vector of predicted Leukocyte telomere length.
}
\description{
A function to calculate the the Leukocyte telomere length (2019)
from a DNA methylation beta value matrix.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
dnamTlO <- dnamTL(hannum_bmiq_m)
}
\references{
Lu AT, Seeboth A, Tsai PC, et al.
DNA methylation-based estimator of telomere length
\emph{Aging} 2019
}
