% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinPACE.R
\name{DunedinPACE}
\alias{DunedinPACE}
\title{Calculate DunedinPACE}
\usage{
DunedinPACE(betas, proportionOfProbesRequired = 0.8)
}
\arguments{
\item{betas}{A numeric matrix of DNA methylation beta values. Rows should represent CpG sites and columns should represent individual samples.}

\item{proportionOfProbesRequired}{A numeric value between 0 and 1 specifying the minimum proportion of required probes that must be present in the input \code{betas} matrix to proceed with the calculation. Defaults to \code{0.8} (80\%). This is automatically adjusted to \code{0.7} for EPICv2 data.}
}
\value{
A named numeric vector containing the calculated DunedinPACE for each sample.
}
\description{
Calculates the DunedinPACE, an epigenetic biomarker that quantifies the pace of biological aging from a single blood sample.
}
\details{
This function calculates DunedinPACE scores from a DNA methylation beta value matrix using a robust pipeline.

The process involves several key steps:
\itemize{
\item \strong{EPICv2 Array Handling}: Automatically detects and preprocesses Illumina EPICv2 array data, adjusting for its specific characteristics.
\item \strong{Probe Overlap Check}: Verifies that a sufficient number of required CpG probes are present in your data.
\item \strong{Missing Data Imputation}: Robustly handles both missing probes and missing values within existing probes.
\item \strong{Quantile Normalization}: Standardizes the data against a gold-standard reference to reduce batch effects and ensure comparability.
\item \strong{Score Calculation}: Computes the final DunedinPACE score using the pre-trained model weights.
}
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
DunedinPACE.out <- DunedinPACE(hannum_bmiq_m)
}
\references{
Belsky DW, Caspi A, Corcoran DL, et al.
DunedinPACE, a DNA methylation biomarker of the pace of aging.
\emph{eLife} 2022
}
