% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunedinPACE.R
\name{dunedinPACE}
\alias{dunedinPACE}
\title{Calculate DunedinPACE}
\usage{
dunedinPACE(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix (Rows: CpGs, Cols: Samples)}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage for
imputation and calculation. Default is 0.5.}

\item{verbose}{Logical. Whether to print progress messages.
Default is \code{TRUE}.}
}
\value{
A named numeric vector containing the calculated DunedinPACE
for each sample.
}
\description{
Calculates the DunedinPACE, an epigenetic biomarker that quantifies the pace of biological aging from a single blood sample.
}
\details{
This function calculates DunedinPACE scores from a DNA methylation beta value matrix using a robust pipeline.

The process involves several key steps:
\itemize{
\item \strong{EPICv2 Array Handling}: Automatically detects and preprocesses Illumina EPICv2 array data, adjusting for its specific characteristics.
\item \strong{Probe Overlap Check}: Verifies that a sufficient number of required CpG probes are present in your data.
\item \strong{Missing Data Imputation}: Robustly handles both missing probes and missing values within existing probes.
\item \strong{Quantile Normalization}: Standardizes the data against a gold-standard reference to reduce batch effects and ensure comparability.
\item \strong{Score Calculation}: Computes the final DunedinPACE score using the pre-trained model weights.
}
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
dunedinPACEOut <- dunedinPACE(hannum_bmiq_m)
}
\references{
Belsky DW, Caspi A, Corcoran DL, et al.
DunedinPACE, a DNA methylation biomarker of the pace of aging.
\emph{eLife} 2022
}
