% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleAge.R
\name{ensembleAge}
\alias{ensembleAge}
\title{Calculate EnsembleAge Epigenetic Clocks}
\usage{
ensembleAge(
  betaM,
  clockVersion = c("HumanMouse", "Static", "Dynamic"),
  minCoverage = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. Sample identifiers in
\code{colnames(betaM)} are recommended but not required. If column names
are absent, temporary internal sample identifiers are generated during
calculation.}

\item{clockVersion}{Character. One of \code{"Dynamic"}, \code{"Static"}, or \code{"HumanMouse"}.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage for
imputation and calculation. Default is 0.5.}

\item{verbose}{Logical. Whether to print progress messages.
Default is \code{TRUE}.}
}
\value{
A \code{list} where each element is a numeric vector of predicted values.
The names of the list elements correspond to the specific sub-clocks
calculated (e.g., \code{HumanMouse_HumanMouse}, \code{Static_Static}, etc.).
}
\description{
Calculates epigenetic age using the EnsembleAge multi-clock framework
(Haghani et al., 2025). This function computes one of the three available
clock versions (Static, Dynamic, or HumanMouse).
}
\note{
\strong{Understanding the \code{clock_version} output:}
\itemize{
\item \strong{\code{"Static"} and \code{"HumanMouse"}:} The returned values are the final,
calibrated age predictors. The \code{"HumanMouse"} clock returns age
normalized by maximum species lifespan.
\item \strong{\code{"Dynamic"}:} This option returns a list of predictors from all
40 clocks in the 'Dynamic' ensemble library. To complete the full
"EnsembleAge.Dynamic" \emph{methodology} as described in the paper,
the user must perform \strong{additional steps}:
\enumerate{
\item Have a dataset with control and intervention groups.
\item Use the returned 40 predictors to find "responsive" clocks
(e.g., those with |Z-score| > 2 between groups).
\item Calculate the final \code{EnsembleAge.Dynamic} value by taking the
\strong{median} of only those responsive clocks.
This function provides the necessary \emph{basis} for this calculation.
}
}
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
# Ensure it's CpGs=rows, Samples=cols
# Calculate the HumanMouse clock version
ensembleAgeOut <- ensembleAge(hannumBmiqM, clockVersion = "HumanMouse")

}
\references{
Haghani, A., Lu, A.T., Yan, Q. et al.
EnsembleAge: enhancing epigenetic age assessment with a
multi-clock framework.
\emph{GeroScience} 2025.
}
