% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_GA.R
\name{epicGa}
\alias{epicGa}
\title{The Gestational Age (GA) clock based on 176 Illumina EPIC CpGs}
\usage{
epicGa(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A matrix of beta values (CpGs in rows, samples in columns).
This matrix must be pre-normalized (e.g., via BMIQ) and imputed.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A named vector of predicted Gestational Ages (in weeks).
}
\description{
The Gestational Age (GA) clock based on 176 Illumina EPIC CpGs
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
epicGaOut <- epicGa(hannumBmiqM, minCoverage = 0)
}
\references{
Haftorn KL, Lee Y, Denault WRP, et al.
An EPIC predictor of gestational age and its application to newborns
conceived by assisted reproductive technologies.
\emph{Clin Epigenetics.} 2021
}
