% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCMIT.R
\name{epiCMIT}
\alias{epiCMIT}
\title{Estimate EpiCMIT scores}
\usage{
epiCMIT(x, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{x}{A numeric matrix, \code{data.frame}, or \code{SummarizedExperiment}
object containing DNA methylation beta values. Rows should be CpG probes and
columns individual samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
A list containing the following entries.
\itemize{
\item hyperScore: EpiCMIT-hyper score of each sample.
\item hypoScore: EpiCMIT-hypo score of each sample.
}
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will
return EpiCMIT scores.
}
\details{
EpiCMIT is based on two groups of CpGs, a 184 age-associated hypermethylated
CpG list and a 1164 hypomethylated CpG list (Duran-Ferrer et al. 2020) .
The EpiCMIT function calculates the average DNAm of the two groups separately
and returns the EpiCMIT_hyper and EpiCMIT_hypo scores.
}
\examples{
data(dnamExample)
beta_matrix <- dnamExample[[1]]

# Example 1: Direct Matrix Input
epicmitOut  <- epiCMIT(x = beta_matrix, verbose = FALSE)

# Example 2: SummarizedExperiment Input
\dontrun{
  library(SummarizedExperiment)
  pheno_data <- dnamExample[[2]]
  rownames(pheno_data) <- colnames(beta_matrix)

  se_obj <- SummarizedExperiment(
    assays = list(beta = beta_matrix),
    colData = pheno_data
  )

  epicmitOut <- epiCMIT(x = se_obj, verbose = FALSE)
}
}
\references{
Duran-Ferrer M, Clot G, Nadeu F, et al.
The proliferative history shapes the DNA methylome of B-cell tumors
and predicts clinical outcome.
\emph{Nat Cancer} 2020
}
