% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCMIT.R
\name{epiCMIT}
\alias{epiCMIT}
\title{Estimate EpiCMIT scores}
\usage{
epiCMIT(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows).}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.5.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
A list containing the following entries.
\itemize{
\item hyperScore: EpiCMIT-hyper score of each sample.
\item hypoScore: EpiCMIT-hypo score of each sample.
}
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will
return EpiCMIT scores.
}
\details{
EpiCMIT is based on two groups of CpGs, a 184 age-associated hypermethylated
CpG list and a 1164 hypomethylated CpG list (Duran-Ferrer et al. 2020) .
The EpiCMIT function calculates the average DNAm of the two groups separately
and returns the EpiCMIT_hyper and EpiCMIT_hypo scores.
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
epicmitOut <- epiCMIT(betaM = lungInvM)
}
\references{
Duran-Ferrer M, Clot G, Nadeu F, et al.
The proliferative history shapes the DNA methylome of B-cell tumors
and predicts clinical outcome.
\emph{Nat Cancer} 2020
}
