% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTOC1.R
\name{epiTOC1}
\alias{epiTOC1}
\title{Estimate epiTOC1 score}
\usage{
epiTOC1(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows).}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
The epiTOC1 score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and
will return the epiTOC1 score.
}
\details{
The epiTOC1 score is calculated as the average beta value of 385 CpGs
described in Yang et al. 2014. These CpGs are promoter CpG sites that
localize to Polycomb group target genes that are unmethylated in 11
different fetal tissue types.
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
epitoc1Out <- epiTOC1(betaM = lungInvM)
}
\references{
Yang Z, Wong A, Kuh D, et al.
Correlation of an epigenetic mitotic clock with cancer risk.
\emph{Genome Biol.} 2016
}
