% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrimAge1.R
\name{GrimAge1}
\alias{GrimAge1}
\title{Calculate GrimAge1}
\usage{
GrimAge1(beta_matrix, age, sex)
}
\arguments{
\item{beta_matrix}{A numeric matrix of DNA methylation beta values. Rows should
represent CpG sites and columns should represent individual samples.}

\item{age}{A numeric vector of chronological ages for the samples corresponding
to the columns in \code{beta_matrix}.}

\item{sex}{A character vector of sample sexes. Must contain "Male" or "Female"
for each sample.}
}
\value{
A data.frame containing the following columns:
\itemize{
\item \code{Sample}: Identifier for each sample.
\item \code{Age}: The input chronological age.
\item \code{Female}: A numeric indicator for sex (1 = Female, 0 = Male).
\item \code{DNAm...}: Columns for each of the predicted surrogate biomarkers (e.g., \code{DNAmADM}, \code{DNAmGDF15}).
\item \code{DNAmGrimAge1}: The final calibrated GrimAge1 score.
}
}
\description{
Calculates DNA methylation GrimAge1, a composite biomarker of mortality risk and biological aging.
}
\details{
This function calculates DNAm GrimAge1 in a multi-step process. First, it predicts DNAm-based surrogate biomarkers for several plasma proteins from the input beta values. These predicted biomarkers, along with chronological age and sex, are then used to calculate a composite mortality risk score. This score is calibrated to the scale of chronological age to produce the final \code{DNAmGrimAge2}.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
age <- PhenoTypesHannum_lv$Age
sex <- ifelse(PhenoTypesHannum_lv$Sex=="F","Female","Male")
GrimAge1.out <- GrimAge1(beta_matrix = hannum_bmiq_m,age,sex)
}
\references{
Lu AT, Quach A, Wilson JG, et al.
DNA methylation GrimAge strongly predicts lifespan and healthspan
\emph{Aging} 2019
}
