% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrimAge2.R
\name{grimAge2}
\alias{grimAge2}
\title{Calculate GrimAge2}
\usage{
grimAge2(betaM, age, sex, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values. Rows should
represent CpG sites and columns should represent individual samples.}

\item{age}{A numeric vector of chronological ages for the samples corresponding
to the columns in \code{betaM}.}

\item{sex}{A character vector of sample sexes. Must contain "Male" or "Female"
for each sample.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A data.frame containing the following columns:
\itemize{
\item \code{SampleID}: Identifier for each sample.
\item \code{DNAm...}: Columns for each of the predicted surrogate biomarkers
(e.g., \code{DNAmADM}, \code{DNAmGDF15}).
\item \code{DNAmGrimAge2}: The final calibrated GrimAge2 score.
}
}
\description{
Calculates DNA methylation GrimAge2, a composite biomarker of mortality risk
and biological aging. This function implements the updated GrimAge2 model.
}
\details{
This function calculates DNAm GrimAge2 in a multi-step process. First, it
predicts DNAm-based surrogate biomarkers for several plasma proteins from
the input beta values. These predicted biomarkers, along with chronological
age and sex, are then used to calculate a composite mortality risk score.
This score is calibrated to the scale of chronological age to produce the
final \code{DNAmGrimAge2}.
}
\examples{
hannumExample <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)
hannumBmiqM <- hannumExample[[1]]
phenoTypesHannum <- hannumExample[[2]]
age <- phenoTypesHannum$Age
sex <- ifelse(phenoTypesHannum$Sex == "F", "Female", "Male")
GrimAge2Oout <- grimAge2(betaM = hannumBmiqM, age, sex)
}
\references{
Lu AT, Binder AM, Zhang J, et al.
DNA methylation GrimAge version 2.
\emph{Aging} 2022
}
