% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hannum.R
\name{hannumClock}
\alias{hannumClock}
\title{Calculate Hannum's DNAm Age (2013)}
\usage{
hannumClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to
the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Hannum epigenetic clock age (2013)
from a DNA methylation beta value matrix.
}
\details{
Implements the Hannum (2013) blood-specific clock. The function calculates
a weighted linear predictor from 71 CpGs found in the input matrix.
This clock is a direct linear model without non-linear transformation.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
hannumClockOut <- hannumClock(hannumBmiqM)

}
\references{
Hannum G, Guinney J, Zhao L, et al.
Genome-wide methylation profiles reveal quantitative views of human aging rates.
\emph{Mol Cell.} 2013
}
