% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath2018.R
\name{horvath2018Clock}
\alias{horvath2018Clock}
\title{Calculate Horvath's Skin & Blood DNAm Age (2018)}
\usage{
horvath2018Clock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector containing one predicted age per
sample, in the same order as the columns of \code{betaM}. If
\code{betaM} has column names, these are retained as the names of the
returned vector; otherwise, an unnamed numeric vector is returned.
}
\description{
A function to calculate the Horvath "Skin & Blood" clock age (2018)
from a DNA methylation beta value matrix.
}
\details{
Implements the Horvath (2013) pan-tissue clock. The function
calculates a weighted linear predictor from 353 CpGs found in the input
matrix and then transforms this value using a non-linear function to
return the final DNAm age.

Implements the Horvath (2018) skin & blood clock. The function calculates
a weighted linear predictor from 391 CpGs and then transforms this value
using the same non-linear function as the Horvath (2013) clock to
return the final DNAm age.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
horvath2018ClockOut <- horvath2018Clock(hannumBmiqM)
}
\references{
Horvath S, Oshima J, Martin GM, et al.
Epigenetic clock for skin and blood cells applied to Hutchinson Gilford
Progeria Syndrome and ex vivo studies.
\emph{Aging} 2018
}
