% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horvath2018.R
\name{horvath2018Clock}
\alias{horvath2018Clock}
\title{Calculate Horvath's Skin & Blood DNAm Age (2018)}
\usage{
horvath2018Clock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding
to the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Horvath "Skin & Blood" clock age (2018)
from a DNA methylation beta value matrix.
}
\details{
Implements the Horvath (2013) pan-tissue clock. The function
calculates a weighted linear predictor from 353 CpGs found in the input
matrix and then transforms this value using a non-linear function to
return the final DNAm age.

Implements the Horvath (2018) skin & blood clock. The function calculates
a weighted linear predictor from 391 CpGs and then transforms this value
using the same non-linear function as the Horvath (2013) clock to
return the final DNAm age.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
horvath2018ClockOut <- horvath2018Clock(hannum_bmiq_m)
}
\references{
Horvath S, Oshima J, Martin GM, et al.
Epigenetic clock for skin and blood cells applied to Hutchinson Gilford
Progeria Syndrome and ex vivo studies.
\emph{Aging} 2018
}
