% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoClock.R
\name{hypoClock}
\alias{hypoClock}
\title{Estimate hypoClock score}
\usage{
hypoClock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
The HypoClock score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will
return the HypoClock score.
}
\details{
The hypoClock score is defined by Teschendorff (2020) and is calculated
as 1 minus the average beta value of 678 solo-WCGW sites.
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
hypoClockOut <- hypoClock(betaM = lungInvM)

}
\references{
Teschendorff AE.
A comparison of epigenetic mitotic-like clocks for cancer risk prediction.
\emph{Genome Med.} 2020
}
