% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoClock.R
\name{hypoClock}
\alias{hypoClock}
\title{Estimate hypoClock score}
\usage{
hypoClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows). Rows
should be Illumina 450k/EPIC CpG identifiers and columns should be samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
The HypoClock score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will
return the HypoClock score.
}
\details{
The hypoClock score is defined by Teschendorff (2020) and is calculated
as 1 minus the average beta value of 678 solo-WCGW sites.
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
hypoClockOut <- hypoClock(betaM = lungInvM)

}
\references{
Teschendorff AE.
A comparison of epigenetic mitotic-like clocks for cancer risk prediction.
\emph{Genome Med.} 2020
}
