% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_Clock.R
\name{icClock}
\alias{icClock}
\title{Calculate the Intrinsic Capacity (IC) Epigenetic Clock.}
\usage{
icClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples. The matrix should not
contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \strong{numeric vector} containing the predicted Intrinsic Capacity (IC) score
for each sample. The vector is named with the sample IDs from the \code{rownames}
of \code{betaM}.
}
\description{
Implements the DNA methylation-based predictor of Intrinsic Capacity,
a biomarker of functional aging and mortality risk.
}
\details{
This function calculates the Intrinsic Capacity (IC) Clock, a novel
DNAm biomarker trained on clinical evaluations of physical and mental
capacities
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
icClockO <- icClock(hannumBmiqM)
}
\references{
Fuentealba M, Rouch L, Guyonnet S, et al.
A blood-based epigenetic clock for intrinsic capacity predicts mortality
and is associated with clinical, immunological and lifestyle factors.
\emph{Nature Aging.} 2025
}
