% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntrinClock.R
\name{intrinClock}
\alias{intrinClock}
\title{IntrinClock Age Prediction}
\usage{
intrinClock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector containing one predicted age per
sample, in the same order as the columns of \code{betaM}. If
\code{betaM} has column names, these are retained as the names of the
returned vector; otherwise, an unnamed numeric vector is returned
}
\description{
Calculates the "IntrinClock" epigenetic age (intrinsic cellular age) using
DNA methylation data(Illumina 450K and EPIC).
This clock is designed to be resistant to changes in immune cell composition.
}
\details{
The IntrinClock utilizes an elastic net model trained on 410 CpGs and
refined to 381 active predictors. It predicts age by calculating a linear
combination of CpG beta values and coefficients, followed by an inverse
Horvath transformation to convert the linear predictor to years.

The model coefficients are stored in \code{IntrinClockCoef}, which contains:
\itemize{
\item Intercept term
\item 381 non-zero CpG coefficients (sparsity optimized)
}
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
intrinClockO <- intrinClock(hannumBmiqM)
}
\references{
Tomusiak, A., Floro, A., Tiwari, R. et al.
Development of an epigenetic clock resistant to changes in
immune cell composition.
\emph{Commun Biol} 2024
}
