% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntrinClock.R
\name{intrinClock}
\alias{intrinClock}
\title{IntrinClock Age Prediction}
\usage{
intrinClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes
and columns should be individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding
to the sample IDs from the input matrix's column names.
}
\description{
Calculates the "IntrinClock" epigenetic age (intrinsic cellular age) using
DNA methylation data(Illumina 450K and EPIC).
This clock is designed to be resistant to changes in immune cell composition.
}
\details{
The IntrinClock utilizes an elastic net model trained on 410 CpGs and
refined to 381 active predictors. It predicts age by calculating a linear
combination of CpG beta values and coefficients, followed by an inverse
Horvath transformation to convert the linear predictor to years.

The model coefficients are stored in \code{IntrinClockCoef}, which contains:
\itemize{
\item Intercept term
\item 381 non-zero CpG coefficients (sparsity optimized)
}
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
intrinClockO <- intrinClock(hannumBmiqM)
}
\references{
Tomusiak, A., Floro, A., Tiwari, R. et al.
Development of an epigenetic clock resistant to changes in
immune cell composition.
\emph{Commun Biol} 2024
}
