% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knight_GA.R
\name{knightGa}
\alias{knightGa}
\title{Calculate the Knight gestational age}
\usage{
knightGa(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A vector of predicted Gestational Ages (in weeks).
}
\description{
A function to calculate the Knight gestational age
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
knightGaOut <- knightGa(hannumBmiqM)

}
\references{
Knight AK, Craig JM, Theda C, et al.
An epigenetic clock for gestational age at birth
based on blood methylation data.
\emph{Genome Biol.} 2016
}
