% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knight_GA.R
\name{knightGa}
\alias{knightGa}
\title{Calculate the Knight gestational age}
\usage{
knightGa(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A matrix of beta values (CpGs in rows, samples in columns).
This matrix must be pre-normalized (e.g., via BMIQ) and imputed.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A named vector of predicted Gestational Ages (in weeks).
}
\description{
A function to calculate the Knight gestational age
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
knightGaOut <- knightGa(hannumBmiqM)
}
\references{
Knight AK, Craig JM, Theda C, et al.
An epigenetic clock for gestational age at birth
based on blood methylation data.
\emph{Genome Biol.} 2016
}
