% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEC.R
\name{leeABEC}
\alias{leeABEC}
\title{Adult Blood-based EPIC Clock (ABEC)}
\usage{
leeABEC(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{betaM}.
}
\description{
Predicts biological age using the Adult Blood-based EPIC Clock (ABEC).
Developed by Lee et al., this model was trained on DNA methylation (DNAm)
data from the Norwegian Mother, Father and Child Cohort Study (MoBa)
(n = 1,592, age range: 19–59 years) using the Illumina EPIC platform.
}
\details{
The function extracts the necessary CpG coefficients from the internal
\code{ABEC_Coef} dataset and applies them to the provided beta value matrix.
It uses an internal helper to handle missing probes and compute the
final age estimates.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
abecOut <- leeABEC(hannumBmiqM)

}
\references{
Lee, Y., Haftorn, K.L., Denault, W.R.P. et al.
Blood-based epigenetic estimators of chronological age in human adults
using DNA methylation data from the Illumina MethylationEPIC array.
\emph{BMC Genomics} 2020
}
