% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEC.R
\name{leeExtendedABEC}
\alias{leeExtendedABEC}
\title{Extended Adult Blood-based EPIC Clock (eABEC)}
\usage{
leeExtendedABEC(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector containing one predicted age per
sample, in the same order as the columns of \code{betaM}. If
\code{betaM} has column names, these are retained as the names of the
returned vector; otherwise, an unnamed numeric vector is returned.
}
\description{
Predicts biological age using the Extended Adult Blood-based EPIC Clock
(eABEC). This model extends the training set of ABEC by incorporating
public data from the Gene Expression Omnibus (GEO), resulting in a
broader age-span (n = 2,227, age range: 18–88 years).
}
\details{
Similar to \code{leeABEC}, this function utilizes the \code{eABEC_Coef}
dataset. It is designed for applications where a wider range of adult
ages is expected.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
eabecOut <- leeExtendedABEC(hannumBmiqM)

}
