% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABEC.R
\name{leeExtendedABEC}
\alias{leeExtendedABEC}
\title{Extended Adult Blood-based EPIC Clock (eABEC)}
\usage{
leeExtendedABEC(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{betaM}.
}
\description{
Predicts biological age using the Extended Adult Blood-based EPIC Clock
(eABEC). This model extends the training set of ABEC by incorporating
public data from the Gene Expression Omnibus (GEO), resulting in a
broader age-span (n = 2,227, age range: 18–88 years).
}
\details{
Similar to \code{leeABEC}, this function utilizes the \code{eABEC_Coef}
dataset. It is designed for applications where a wider range of adult
ages is expected.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
eABEC_o <- leeExtendedABEC(hannum_bmiq_m)

}
