% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mayne_GA.R
\name{mayneGa}
\alias{mayneGa}
\title{Calculate the Mayne Placental Gestational Age.}
\usage{
mayneGa(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \strong{numeric vector} containing the predicted gestational age (in weeks) for
each sample.
}
\description{
Implements the 62-CpG placental epigenetic clock for estimating gestational
age (GA), as described by Mayne et al. (2017).
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
mayneGaOut <- mayneGa(hannumBmiqM)
}
\references{
Mayne BT, Leemaqz SY, Smith AK, Breen J, Roberts CT, Bianco-Miotto T.
Accelerated placental aging in early onset preeclampsia pregnancies
identified by DNA methylation.
\emph{Epigenomics} 2017
}
