% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McCartney_Trait.R
\name{mcCartneyTrait}
\alias{mcCartneyTrait}
\title{Calculate Epigenetic Surrogates for Complex Traits (McCartney et al. 2018)}
\usage{
mcCartneyTrait(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values. Rows should be CpG probes and
columns should be individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \code{list} containing 10 named elements, one for each trait. Each element is
a \strong{named numeric vector} of the calculated epigenetic scores.
\itemize{
\item \strong{\code{BMI}}: Numeric vector of predicted scores.
\item \strong{\code{Smoking}}: Numeric vector of predicted scores.
\item \strong{\code{Alcohol}}: Numeric vector of predicted scores.
\item \strong{\code{Education}}: Numeric vector of predicted scores.
\item \strong{\code{Total_cholesterol}}: Numeric vector of predicted scores.
\item \strong{\code{HDL_cholesterol}}: Numeric vector of predicted scores.
\item \strong{\code{LDL_cholesterol}}: Numeric vector of predicted scores.
\item \strong{\code{Total_HDL_ratio}}: Numeric vector of predicted scores.
\item \strong{\code{WHR}}: Numeric vector of predicted scores.
\item \strong{\code{Body_fat_Perc}}: Numeric vector of predicted scores.
}
Each vector is named with the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Implements the elastic net (EN) predictors for 10 complex traits and
biomarkers using DNAm data, as described by McCartney et al. (2018).
}
\details{
This function computes epigenetic surrogate scores for 10 different traits
based on models trained on blood methylation data.

The function iterates through and calculates scores for the following traits:
\itemize{
\item \code{BMI} (Body Mass Index)
\item \code{Smoking} (Smoking pack-years)
\item \code{Alcohol} (Alcohol consumption units per week)
\item \code{Education} (Years of education)
\item \code{Total_cholesterol}
\item \code{HDL_cholesterol}
\item \code{LDL_cholesterol}
\item \code{Total_HDL_ratio} (Ratio of Total to HDL cholesterol)
\item \code{WHR} (Waist-to-Hip Ratio)
\item \code{Body_fat_Perc} (Body Fat Percentage)
}
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
mcCartneyTraitOut <- mcCartneyTrait(hannumBmiqM)

}
\references{
McCartney DL, Hillary RF, Stevenson AJ, et al.
Epigenetic prediction of complex traits and death.
\emph{Genome Biol.} 2018
}
