% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeerMouse.R
\name{meerMouse}
\alias{meerMouse}
\title{Meer Whole Lifespan Multi-Tissue Mouse DNA Methylation Age Predictor}
\usage{
meerMouse(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of beta values (ranging from 0 to 1). Rows must
be labeled with mouse genomic coordinates (typically in 'chr:pos' format
based on the mm10/GRCm38 assembly) corresponding to the clock's CpG sites.
Columns represent individual samples.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted chronological ages in months
}
\description{
Predicts the chronological age of mouse samples based on the Whole Lifespan
Multi-tissue (WLMT) DNA methylation clock developed by Meer et al. (2018).
}
\details{
The clock is a robust age predictor that covers the entire mouse lifespan
and remains unbiased across different age groups and sexes. It utilizes an
elastic net regression model across 435 specific CpG sites. The raw output
is calculated in days and is converted to months (divided by 30.5) by this function.
}
\examples{
exampleMat <- loadOmniAgeRdata(
    "omniager_mouse_rrbs_example",
    verbose = FALSE
)
predRes <- meerMouse(exampleMat)

}
\references{
Meer, Margarita V et al.
A whole lifespan mouse multi-tissue DNA methylation clock.
\emph{Elife} 2018
}
