% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanMammalianBlood.R
\name{panMammalianBlood}
\alias{panMammalianBlood}
\title{Calculate Universal Pan-Mammalian Blood Epigenetic Clocks}
\usage{
panMammalianBlood(
  betaM,
  speciesName,
  anageData = NULL,
  minCoverage = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{betaM}{A numeric matrix (Rows: CpGs, Cols: Samples).}

\item{speciesName}{A character string or vector specifying the Latin species
name(s) (e.g., "Homo sapiens").}

\item{anageData}{A data.frame containing the AnAge database information.
Must include 'SpeciesLatinName', 'GestationTimeInYears',
'averagedMaturity.yrs', and 'maxAge'.}

\item{minCoverage}{Numeric (0-1). Minimum required proportion of CpGs present.
Default is 0.5.}

\item{verbose}{Logical. Whether to print status messages.}
}
\value{
A data.frame containing the 'Sample', 'SpeciesLatinName',
and the calculated ages: 'DNAmRelativeAge',
'DNAmAgePanMammalianBlood2', 'DNAmRelativeAdultAge' and
'DNAmAgePanMammalianBlood3'. Any additional columns
from the input \code{sampleInfo} (like 'Age', 'Tissue') will also be returned.
}
\description{
Applies the two universal pan-mammalian BLOOD clocks (Clock 2 and 3)
from Lu et al. (2023) to a given dataset of DNA methylation values.
}
\details{
This function is a specific adaptation for the Blood-only clocks.
It requires the 'PanMammalianBloodCoef' data object to be available
in the R package's data directory.
}
\examples{

downloadOmniAgeRExample("Tursiops_example")
loadOmniAgeRExample("Tursiops_example")

## This anage_data is from https://github.com/shorvath/MammalianMethylationConsortium
data("anage_data")

# Run the calculation with progress messages

clockResults <- panMammalianBlood(
 betaM = example_beta_m,
 speciesName = example_sample_info$SpeciesLatinName,
 anageData = anage_data
)
}
\references{
Lu, A.T., Fei, Z., Haghani, A. et al.
Universal DNA methylation age across mammalian tissues.
\emph{Nat Aging.} 2023
}
