% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanMammalianSkin.R
\name{panMammalianSkin}
\alias{panMammalianSkin}
\title{Calculate Universal Pan-Mammalian Skin Epigenetic Clocks}
\usage{
panMammalianSkin(
  betaM,
  speciesName,
  anageData = NULL,
  minCoverage = 0,
  verbose = TRUE
)
}
\arguments{
\item{betaM}{A numeric matrix (Rows: CpGs, Cols: Samples).}

\item{speciesName}{A character string or vector specifying the Latin species
name(s) (e.g., "Homo sapiens").}

\item{anageData}{A data.frame containing the AnAge database information.
Must include 'SpeciesLatinName', 'GestationTimeInYears',
'averagedMaturity.yrs', and 'maxAge'.}

\item{minCoverage}{Numeric (0-1). Minimum required proportion of CpGs present.
Default is 0.}

\item{verbose}{Logical. Whether to print status messages.}
}
\value{
A data.frame containing the 'Sample', 'SpeciesLatinName',
and the calculated ages: 'DNAmAgePanMammalianSkin2',
'DNAmAgePanMammalianSkin3', 'DNAmRelativeAge_Skin', and
'DNAmRelativeAdultAge_Skin'.
}
\description{
Applies the two universal pan-mammalian BLOOD clocks (Clock 2 and 3)
from Lu et al. (2023) to a given dataset of DNA methylation values.
}
\details{
This function is a specific adaptation for the Skin-only clocks.
It requires the 'PanMammalianSkinCoef' data object to be available
in the R package's data directory.
}
\examples{
tursiopsExample <- loadOmniAgeRdata(
    "omniager_tursiops_example",
    verbose = FALSE
)

## This anage_data is from
## https://github.com/shorvath/MammalianMethylationConsortium
anageData <- loadOmniAgeRdata(
    "omniager_anage_data",
    verbose = FALSE
)

# Run the calculation with progress messages
clockResults <- panMammalianSkin(
    betaM = tursiopsExample$beta_m,
    speciesName = tursiopsExample$PhenoTypes$SpeciesLatinName,
    anageData = anageData
)

}
\references{
Lu, A.T., Fei, Z., Haghani, A. et al.
Universal DNA methylation age across mammalian tissues.
\emph{Nat Aging.} 2023
}
