% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASTA.R
\name{pastaScores}
\alias{pastaScores}
\title{Calculate PASTA-based transcriptomic age scores}
\usage{
pastaScores(
  x,
  filterGenes = TRUE,
  rankNorm = TRUE,
  reg = FALSE,
  pasta = TRUE,
  ct46 = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A numeric matrix, \code{data.frame}, or \code{SummarizedExperiment}
object containing an expression matrix (numeric) with \strong{genes as rows} and
\strong{samples as columns}}

\item{filterGenes}{Logical. If \code{TRUE} (default), the matrix is
subsetted to retain only the genes utilized by the pre-trained models.}

\item{rankNorm}{Logical. If \code{TRUE} (default), applies a rank-based
inverse normal transformation (rank-normalization) to the expression data.}

\item{reg}{Logical. If \code{TRUE}, computes the REG (regression)
age score.}

\item{pasta}{Logical. If \code{TRUE} (default), computes the PASTA age score.}

\item{ct46}{Logical. If \code{TRUE}, computes the CT46 age score.}

\item{verbose}{Logical. Whether to print status messages.
Default is \code{TRUE}.}
}
\value{
A \code{list} where each element is a numeric vector of predicted age scores
for the requested model(s) (e.g., \code{res_list$PASTA}, \code{res_list$REG}).
}
\description{
This function computes transcriptomic age acceleration scores based on the
models described by Salignon et al. (2025), including the primary PASTA
score, a standard regression (REG) score, and the CT46 score.
}
\examples{
library(glmnet)
data("ScPbmcExample")
expr_mat <- as.matrix(SummarizedExperiment::assay(ScPbmcExample, "logcounts"))
cell_meta <- as.data.frame(SummarizedExperiment::colData(ScPbmcExample))
groups <- paste(cell_meta$donor_id, cell_meta$celltype, sep = "_")
unique_groups <- unique(groups)
 pseudobulk_data_list <- lapply(unique_groups, function(grp) {
 rowSums(expr_mat[, groups == grp, drop = FALSE])
})

pseudobulk_data_matrix <- do.call(cbind, pseudobulk_data_list)
colnames(pseudobulk_data_matrix) <- unique_groups

pastaRes <- pastaScores(pseudobulk_data_matrix, filterGenes = TRUE, 
                        rankNorm = TRUE)


}
\references{
Salignon, J., Tsiokou, M., Marqués, P. et al.
Pasta, an age-shift transcriptomic clock, maps the chemical and genetic determinants of aging and rejuvenation.
\emph{bioRxiv.} 2025
}
