% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASTA.R
\name{pastaScores}
\alias{pastaScores}
\title{Calculate PASTA-based transcriptomic age scores}
\usage{
pastaScores(
  mat,
  filterGenes = TRUE,
  rankNorm = TRUE,
  reg = TRUE,
  pasta = TRUE,
  ct46 = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{mat}{An expression matrix (numeric) with \strong{genes as rows} and
\strong{samples as columns}.}

\item{filterGenes}{Logical. If \code{TRUE} (default), the matrix is
subsetted to retain only the genes utilized by the pre-trained models.}

\item{rankNorm}{Logical. If \code{TRUE} (default), applies a rank-based
inverse normal transformation (rank-normalization) to the expression data.}

\item{reg}{Logical. If \code{TRUE} (default), computes the REG (regression)
age score.}

\item{pasta}{Logical. If \code{TRUE} (default), computes the PASTA age score.}

\item{ct46}{Logical. If \code{TRUE} (default), computes the CT46 age score.}

\item{verbose}{Logical. Whether to print status messages.
Default is \code{TRUE}.}
}
\value{
A \code{list} where each element is a numeric vector of predicted age scores
for the requested model(s) (e.g., \code{res_list$PASTA}, \code{res_list$REG}).
}
\description{
This function computes transcriptomic age acceleration scores based on the
models described by Salignon et al. (2025), including the primary PASTA
score, a standard regression (REG) score, and the CT46 score.
}
\examples{
# 1. Fast runnable example
print("Ready to calculate PASTA scores.")

\donttest{
library(magrittr)
library(Seurat)
library(glmnet)

seu <- loadOmniAgeRdata(
    "omniager_seu_gabitto_2024_filtered",
    verbose = FALSE
)

seu$age <- seu$development_stage \%>\%
    gsub("-year.*", "", .) \%>\%
    gsub("-", " ", .) \%>\%
    gsub("80 year old and over stage", "85", .)

seuBulk <- makePseudobulksPasta(
    seu,
    poolBy = c("cell_type", "age"),
    chunkSize = 512,
    verbose = FALSE
)

lognormMatrix <- GetAssayData(seuBulk, assay = "RNA", layer = "data")
lognormMatrix <- as.matrix(lognormMatrix)

seuBulkMeta <- seuBulk[[c("chunkSize", "cell_type", "age")]]
seuBulkMeta$age <- as.numeric(seuBulkMeta$age)
pastaRes <- pastaScores(lognormMatrix, filterGenes = TRUE, rankNorm = TRUE)
}

}
\references{
Salignon, J., Tsiokou, M., Marqués, P. et al.
Pasta, an age-shift transcriptomic clock, maps the chemical and genetic determinants of aging and rejuvenation.
\emph{bioRxiv.} 2025
}
