% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCClocks.R
\name{pcClocks}
\alias{pcClocks}
\title{Calculate Principal Component (PC) Epigenetic Clocks}
\usage{
pcClocks(x, age, sex, clockData, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{x}{A numeric matrix, \code{data.frame}, or \code{SummarizedExperiment}
object containing DNA methylation beta values. Rows should be CpG probes and
columns individual samples.}

\item{age}{A numeric vector of chronological ages for each sample,
in the same order as the columns of \code{DNAm}.}

\item{sex}{A character vector of biological sex for each sample, in the
same order as the columns of \code{DNAm}. Values of "Female" are
encoded as 1; all other values are encoded as 0.}

\item{clockData}{The pre-loaded data object from
\code{loadOmniAgeRData("PCClocks_data")}.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.}

\item{verbose}{Logical. Whether to print status messages.}
}
\value{
A data.frame containing the original \code{SampleID} columns,
appended with 14 new columns for the calculated PC clock values
(e.g., \code{PCHorvath2013}, \code{PCHannum}, \code{PCGrimAge1}, etc.).
}
\description{
Calculates a suite of Principal Component (PC)-based epigenetic clocks
based on the methodology from Higgins-Chen et al. (2022).

This function computes PC-based versions of Horvath2013, Horvath2018, Hannum,
PhenoAge, and GrimAge1, along with their principal components.
}
\examples{
\dontrun{
# ====================================================================
# Example 1: Direct Matrix Input
# ====================================================================
  hannumExample <- OmniAgeRData::getOmniAgeRData(
      "omniager_hannum_example",
      verbose = FALSE
  )
  
  pcClockData <- OmniAgeRData::getOmniAgeRData(
      "PCClocks_data",
      verbose = FALSE
  )
  
  hannumBmiqM <- hannumExample[[1]]
  phenoTypesHannum <- hannumExample[[2]]
  age <- phenoTypesHannum$Age
  sex <- ifelse(phenoTypesHannum$Sex == "F", "Female", "Male")
  pcClocksOut <- pcClocks(hannumBmiqM, age, sex, pcClockData)
  
# ====================================================================
# Example 2: SummarizedExperiment Input
# ====================================================================
  library(SummarizedExperiment)

  se_obj <- SummarizedExperiment(
    assays = list(beta = hannumBmiqM),
    colData = hannumExample[[2]]
  )

  pcClocksOut <-  pcClocks(se_obj,age, sex, pcClockData)
}

}
\references{
Higgins-Chen AT, Thrush KL, Wang Y, et al.
A computational solution for bolstering reliability of epigenetic clocks:
Implications for clinical trials and longitudinal tracking.
\emph{Nat Aging.} (2022).
}
