% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCClocks.R
\name{pcClocks}
\alias{pcClocks}
\title{Calculate Principal Component (PC) Epigenetic Clocks}
\usage{
pcClocks(betaM, age, sex, clockData, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation (beta) values.
\strong{Samples must be in columns} and CpG probes in rows.}

\item{age}{A numeric vector of chronological ages for each sample,
in the same order as the columns of \code{DNAm}.}

\item{sex}{A character vector of biological sex for each sample, in the
same order as the columns of \code{DNAm}. Values of "Female" are
encoded as 1; all other values are encoded as 0.}

\item{clockData}{The pre-loaded data object from
\code{loadOmniAgeRData("PCClocks_data")}.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.5.}

\item{verbose}{Logical. Whether to print status messages.}
}
\value{
A data.frame containing the original \code{Sample_ID}, \code{Age}, and \code{Female}
columns, appended with 14 new columns for the calculated PC clock values
(e.g., \code{PCHorvath2013}, \code{PCHannum}, \code{PCGrimAge1}, etc.).
}
\description{
Calculates a suite of Principal Component (PC)-based epigenetic clocks
based on the methodology from Higgins-Chen et al. (2022).

This function computes PC-based versions of Horvath2013, Horvath2018, Hannum,
PhenoAge, and GrimAge1, along with their principal components.
}
\examples{

# Download the external data
download_OmniAgeR_data(clocks = "PCClocks") #  ZENODO_DOI: "10.5281/zenodo.17162604"

# Either path to the data
RData <- getOmniAgeRPath()
# OR
RData <- load_OmniAgeR_data(object_name = "PCClocks_data")


downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
age <- PhenoTypesHannum_lv$Age
sex <- ifelse(PhenoTypesHannum_lv$Sex == "F", "Female", "Male")
pcClocksOut <- pcClocks(hannum_bmiq_m, age, sex, RData)

}
\references{
Higgins-Chen AT, Thrush KL, Wang Y, et al.
A computational solution for bolstering reliability of epigenetic clocks:
Implications for clinical trials and longitudinal tracking.
\emph{Nat Aging.} (2022).
}
