% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedBE.R
\name{pedBEClock}
\alias{pedBEClock}
\title{The PedBE (Pediatric Buccal) Clock for DNAm Age in Children}
\usage{
pedBEClock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A matrix of beta values (CpGs in rows, samples in columns).
This matrix must be pre-normalized (e.g., via BMIQ) and imputed.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \strong{numeric vector} containing the predicted DNAm age
for each sample. The vector is named with the sample IDs from the \code{rownames}
of \code{betaM}.
}
\description{
Implements the Pediatric Buccal Epigenetic (PedBE) clock, specifically
developed to estimate DNA methylation age in
\strong{pediatric (childhood) samples}, as described by McEwen et al. (2020).
}
\details{
This clock is specifically trained on and designed for pediatric buccal
epithelial (cheek swab) samples. The calculation is a two-step process:
\enumerate{
\item A linear predictor is first calculated from the beta values using the
94-CpG elastic net coefficients . This value represents a \emph{transformed} age.
\item A non-linear inverse age transformation (via the transformation function
from Horvath, 2013) is then applied. This converts the transformed age into
a final estimate of chronological age in years.
}
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
pedBEClockOut <- pedBEClock(hannumBmiqM)

}
\references{
McEwen LM, O'Donnell KJ, McGill MG, et al.
The PedBE clock accurately estimates DNA methylation age in
pediatric buccal cells.
\emph{Proc Natl Acad Sci U S A.} 2020
}
