% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoAge.R
\name{PhenoAge}
\alias{PhenoAge}
\title{Calculate DNAm PhenoAge}
\usage{
PhenoAge(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values. Rows should represent CpG sites and columns should represent individual samples.}
}
\value{
A named numeric vector containing the calculated PhenoAge for each sample.
}
\description{
CCalculates the DNA methylation PhenoAge, an epigenetic biomarker of aging, based on a matrix of DNA methylation beta values. The function implements the model originally developed by Levine et al. (2018), which uses a weighted linear combination of 513 specific CpG sites to predict phenotypic age.
}
\details{
This function calculates DNAm PhenoAge based on the model by Levine et al. (2018), which predicts phenotypic age by calculating a weighted sum of the beta values from 513 specific CpG sites.
The function automatically loads the required model coefficients, matches them with the CpGs in the input \code{beta.m} matrix, and computes the age.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
PhenoAge.out <- PhenoAge(hannum_bmiq_m)
}
\references{
Levine ME, Lu AT, Quach A, et al.
An epigenetic biomarker of aging for lifespan and healthspan.
\emph{Aging} 2018
}
