% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipekClock.R
\name{pipekElasticNet}
\alias{pipekElasticNet}
\title{Pipek's Multi-tissue Elastic Net Epigenetic Clock (239 CpGs)}
\usage{
pipekElasticNet(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector containing one predicted age per
sample, in the same order as the columns of \code{betaM}. If
\code{betaM} has column names, these are retained as the names of the
returned vector; otherwise, an unnamed numeric vector is returned.
}
\description{
Implements the "elasticNet (239)" epigenetic clock model proposed by
Pipek and Csabai (2023). This model was trained using elastic net
regression on a large multi-tissue dataset containing methylation
data from Illumina 27K, 450K, and EPIC platforms. It is designed to
provide improved accuracy on EPIC array data compared to the original
Horvath2013 clock.
}
\details{
Implements the "elasticNet (239)" model using \strong{239 CpGs} shared
across Illumina 27K, 450K, and EPIC arrays.

\strong{Input Requirements:}
\itemize{
\item \strong{Complete Data:} Missing values must be imputed prior to input.
\item \strong{Normalization:} Recommended but not strictly required.
}

Internally applies Horvath's (2013) log-linear age transformation.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
pipekElasticNetOut <- pipekElasticNet(hannumBmiqM)

}
\references{
Pipek, O.A., Csabai, I.
A revised multi-tissue, multi-platform epigenetic clock model for
methylation array data. \emph{J Math Chem} 2023
}
