% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipekClock.R
\name{pipekElasticNet}
\alias{pipekElasticNet}
\title{Pipek's Multi-tissue Elastic Net Epigenetic Clock (239 CpGs)}
\usage{
pipekElasticNet(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{betaM}.
}
\description{
Implements the "elasticNet (239)" epigenetic clock model proposed by
Pipek and Csabai (2023). This model was trained using elastic net
regression on a large multi-tissue dataset containing methylation
data from Illumina 27K, 450K, and EPIC platforms. It is designed to
provide improved accuracy on EPIC array data compared to the original
Horvath2013 clock.
}
\details{
Implements the "elasticNet (239)" model using \strong{239 CpGs} shared
across Illumina 27K, 450K, and EPIC arrays.

\strong{Input Requirements:}
\itemize{
\item \strong{Complete Data:} Missing values must be imputed prior to input.
\item \strong{Normalization:} Recommended but not strictly required.
}

Internally applies Horvath's (2013) log-linear age transformation.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
pipekElasticNetOut <- pipekElasticNet(hannum_bmiq_m)

}
\references{
Pipek, O.A., Csabai, I.
A revised multi-tissue, multi-platform epigenetic clock model for
methylation array data. \emph{J Math Chem} 2023
}
