% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipekClock.R
\name{PipekRetrainedh}
\alias{PipekRetrainedh}
\title{Pipek's Retrained Horvath Epigenetic Clock (308 CpGs)}
\usage{
PipekRetrainedh(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values.
\strong{Rows must correspond to samples and columns to CpGs.}
\code{rownames} (sample IDs) and \code{colnames} (CpG probe IDs) are required.
The matrix should not contain \code{NA} values.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Implements the "retrained H (308)" epigenetic clock model proposed by Pipek and Csabai (2023).
This model is a direct recalibration of the original Horvath clock probes using a large
multi-platform training set.
}
\details{
Implements the "retrained H (308)" model. It includes all \strong{308 CpG sites} from the original
Horvath pan-tissue clock that are present across 27K, 450K, and EPIC platforms.

\strong{Key Features:}
\itemize{
\item \strong{No Feature Selection:} Coefficients were simply refitted to the new data without dropping probes.
\item \strong{Robust Update:} Serves as a direct update to the Horvath clock to correct for accuracy loss on EPIC arrays.
}

\strong{Input Requirements:}
\itemize{
\item \strong{Complete Data:} Missing values must be imputed.
\item \strong{Normalization:} Recommended.
}

Internally applies Horvath's (2013) log-linear age transformation.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
PipekRetrainedh_o <- PipekRetrainedh(hannum_bmiq_m)

}
\references{
Pipek, O.A., Csabai, I.
A revised multi-tissue, multi-platform epigenetic clock model for methylation array data.
\emph{J Math Chem} 2023
}
