% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retro_age.R
\name{retroAge}
\alias{retroAge}
\title{Calculate the Retro-age Epigenetic Clock}
\usage{
retroAge(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A list containing the predicted age for the "V1" and "V2" clocks.
}
\description{
Calculates the "Retro-age," a retroelement-based epigenetic clock for
chronological age, based on the models developed by Ndhlovu et al. (2024).
This function computes both Version 1 (V1) and Version 2 (V2) of the clock.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
retroAgeRes <- retroAge(hannumBmiqM)
}
\references{
Ndhlovu LC, Bendall ML, Dwaraka V, et al.
Retro-age: A unique epigenetic biomarker of aging captured by DNA methylation states of retroelements.
\emph{Aging Cell.} 2024
}
