% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retro_age.R
\name{retroAge}
\alias{retroAge}
\title{Calculate the Retro-age Epigenetic Clock}
\usage{
retroAge(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value between 0 and 1 (default is 0.5).
Specifies the minimum proportion of required CpGs that must be present
in the input matrix for the clock calculation to proceed.}

\item{verbose}{A logical value. If TRUE (default), the function will
print messages detailing the calculation steps.}
}
\value{
A list containing the predicted age for the "V1" and "V2" clocks.
}
\description{
Calculates the "Retro-age," a retroelement-based epigenetic clock for
chronological age, based on the models developed by Ndhlovu et al. (2024).
This function computes both Version 1 (V1) and Version 2 (V2) of the clock.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
Retro_age.o <- Retro_age(hannum_bmiq_m)
}
\references{
Ndhlovu LC, Bendall ML, Dwaraka V, et al.
Retro-age: A unique epigenetic biomarker of aging captured by DNA methylation states of retroelements.
\emph{Aging Cell.} 2024
}
