% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scImmuAging.R
\name{scImmuAging}
\alias{scImmuAging}
\title{Run the scImmuAging prediction pipeline for multiple cell types}
\usage{
scImmuAging(seurat_object, cell_types)
}
\arguments{
\item{seurat_object}{A Seurat object. The meta.data must contain "donor_id", "age" and "celltype" columns.}

\item{cell_types}{A character vector specifying the cell types for which to predict age.Example: \code{c("CD4T", "CD8T", "MONO")}. Valid types are "CD4T", "CD8T", "MONO", "NK", "B".}
}
\value{
A list where each element is named by a cell type from the input vector.
Each of these elements is itself a list containing two data frames:
\describe{
\item{BootstrapCell}{A data frame with age predictions for each bootstrapped pseudocell,
containing the columns \code{donor_id}, \code{age}, and \code{Prediction}.}
\item{Donor}{A data frame with the final aggregated age prediction for each donor,
containing the columns \code{donor_id}, \code{age}, and \code{predicted}.}
}
}
\description{
This function takes as input a log-normalized matrix of scRNA-seq data and will return cell type-specific predicted age.
}
\details{
This function takes a Seurat object, preprocesses the scRNA-seq data for one or more specified cell types, and predicts age using the pre-trained scImmuAging models. It iterates through a vector of cell types, performs the full analysis for each, and returns a nested list containing all results.
}
\examples{
library(Seurat)
library(glmnet)
download_OmniAgeR_example("Yazar_CD4T_CD8T_example")
load_OmniAgeR_example("Yazar_CD4T_CD8T_example")
scImmuAging.out <- scImmuAging(seurat_obj,c("CD4T","CD8T"))

}
\references{
Li W, Zhang Z, Kumar S, et al.
Single-cell immune aging clocks reveal inter-individual heterogeneity during infection and vaccination.
\emph{Nat Aging} 2025
}
