% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_scImmuAging.R
\name{scImmuAgingPreProcess}
\alias{scImmuAgingPreProcess}
\title{Pre-Processing of input dataset for scImmuAging}
\usage{
scImmuAgingPreProcess(inputObj, cellType, model, markerGenes)
}
\arguments{
\item{inputObj}{A Seurat object. Meta data must contain "donorId" and
"age" columns.}

\item{cellType}{A string specifying the cell type.}

\item{model}{The pre-trained model object.}

\item{markerGenes}{Character vector of selected marker genes.}
}
\value{
A nested and unnested data frame with pseudocells.
}
\description{
Pre-Processing of input dataset for scImmuAging
}
