% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemTOC.R
\name{stemTOC}
\alias{stemTOC}
\title{Estimate stemTOC score}
\usage{
stemTOC(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric DNA methylation beta-value matrix with CpG probe
identifiers as row names and samples as columns. CpG identifiers in
\code{rownames(betaM)} are required. Sample identifiers in
\code{colnames(betaM)} are optional. The matched CpG values used for
calculation must not contain missing values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
The stemTOC score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and
will return the stemTOC score.
}
\details{
The function will return the 0.95 upper quantile of the 371 stemTOC CpGs.
Compared to stemTOCvitro CpGs, the stemTOC CpGs are filtered for
significant DNA hypermethylation with chronological age
in large in-vivo datasets
}
\examples{
lungInv <- loadOmniAgeRdata(
    "omniager_lung_inv",
    verbose = FALSE
)
lungInvM <- lungInv$bmiq_m
stemtocOut <- stemTOC(betaM = lungInvM)

}
\references{
Zhu, T., Tong, H., Du, Z. et al.
An improved epigenetic counter to track mitotic age in normal
and precancerous tissues.
\emph{Nat Commun} 2024
}
