% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemTOC.R
\name{stemTOC}
\alias{stemTOC}
\title{Estimate stemTOC score}
\usage{
stemTOC(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
The stemTOC score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return the stemTOC score.
}
\details{
The function will return the 0.95 upper quantile of the 371 stemTOC CpGs. Compared to stemTOCvitro CpGs, the stemTOC CpGs are filtered for significant DNA hypermethylation with chronological age in large in-vivo datasets
}
\examples{
download_OmniAgeR_example("LungInv")
load_OmniAgeR_example("LungInv")
stemTOC.v<-stemTOC(data.m = bmiq.m)

}
\references{
Zhu, T., Tong, H., Du, Z. et al.
An improved epigenetic counter to track mitotic age in normal and precancerous tissues.
\emph{Nat Commun} 2024
}
