% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemTOCvitro.R
\name{stemTOCvitro}
\alias{stemTOCvitro}
\title{Estimate stemTOCvitro score}
\usage{
stemTOCvitro(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
The stemTOCvitro score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return the stemTOCvitro score.
}
\details{
The function will return the 0.95 upper quantile of the 629 stemTOCvitro CpGs.
}
\examples{
download_OmniAgeR_example("LungInv")
load_OmniAgeR_example("LungInv")
stemTOCvitro.v<-stemTOCvitro(data.m = bmiq.m)

}
\references{
Zhu, T., Tong, H., Du, Z. et al.
An improved epigenetic counter to track mitotic age in normal and precancerous tissues.
\emph{Nat Commun} 2024
}
