% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemTOCvitro.R
\name{stemTOCvitro}
\alias{stemTOCvitro}
\title{Estimate stemTOCvitro score}
\usage{
stemTOCvitro(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNAm beta values (probes as rows). Rows
should be Illumina 450k/EPIC CpG identifiers and columns should be samples.}

\item{minCoverage}{Numeric (0-1). Minimum required probe coverage.
Default is 0.5.}

\item{verbose}{Logical. Whether to print coverage statistics.}
}
\value{
The stemTOCvitro score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix
and will return the stemTOCvitro score.
}
\details{
The function will return the 0.95 upper quantile of
the 629 stemTOCvitro CpGs.
}
\examples{
downloadOmniAgeRExample("LungInv")
loadOmniAgeRExample("LungInv")
stemTOCvitroOut <- stemTOCvitro(data.m = bmiq.m)

}
\references{
Zhu, T., Tong, H., Du, Z. et al.
An improved epigenetic counter to track mitotic age in normal
and precancerous tissues.
\emph{Nat Commun} 2024
}
