% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemsAge.R
\name{SystemsAge}
\alias{SystemsAge}
\title{Calculate Systems Age and 11 System-Specific Scores}
\usage{
SystemsAge(DNAm, RData = NULL)
}
\arguments{
\item{DNAm}{A numeric matrix or data.frame of DNA methylation beta values.
\strong{Rows must correspond to CpGs} and \strong{columns to samples}.}

\item{RData}{A required parameter specifying the location or content of the
pre-trained 'SystemsAge_data' object. This can be either:
\itemize{
\item{1. A \strong{character string} path to the directory containing the
'SystemsAge_data.qs' file (e.g., from \code{get_OmniAgeR_path()}).}
\item{2. The pre-loaded \strong{list} object itself, loaded via
\code{load_OmniAgeR_data(object_name = "SystemsAge_data")}.}
}}
}
\value{
A data.frame where the first column is \code{Sample_ID} (derived from the
\code{colnames} of the input \code{DNAm} matrix) and the subsequent 13 columns
contain the calculated scores. These include the 11 system scores
(e.g., 'Blood', 'Brain', 'Heart'), the 'Age_prediction' score,
and the composite 'SystemsAge' score. All scores are scaled to the
unit of years.
}
\description{
Implements the Systems Age epigenetic clock, a multi-system aging biomarker
based on DNA methylation (DNAm). This function calculates a composite 'Systems Age'
as well as 11 individual physiological system aging scores from a single
blood methylation dataset.
}
\details{
The Systems Age framework, developed by Sehgal et al. (2025), quantifies
aging heterogeneity across 11 distinct physiological systems. It integrates
supervised and unsupervised machine learning with clinical biomarkers and
mortality risk to derive system-specific scores.

This function applies the pre-trained models to new methylation data. It
requires the \code{SystemsAge_data.qs} object, which contains the necessary
PCA models, elastic net coefficients, and scaling parameters from the
original study.

This data object must first be downloaded using
\code{download_OmniAgeR_data(clocks = "SystemsAge")}.

For computational efficiency, especially when processing multiple datasets
or running the function in a loop, it is recommended to load the data
object once using \code{load_OmniAgeR_data(object_name = "SystemsAge_data")}
and pass the resulting list to the \code{RData} parameter.
}
\examples{
# Download the external data
download_OmniAgeR_data(clocks = "SystemsAge") #  ZENODO_DOI: "10.5281/zenodo.17162604"

# Either path to the data
RData <- get_OmniAgeR_path()
# OR
RData <- load_OmniAgeR_data(object_name = "SystemsAge_data")


download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
SystemsAge.o <- SystemsAge(hannum_bmiq_m,RData = RData)


}
\references{
Sehgal, R., Markov, Y., Qin, C. et al.
Systems Age: a single blood methylation test to quantify aging heterogeneity across 11 physiological systems.
\emph{Nat Aging} (2025).
}
