% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniversalPanMammalianClocks.R
\name{universalPanMammalianClocks}
\alias{universalPanMammalianClocks}
\title{Calculate Universal Pan-Mammalian Epigenetic Clocks}
\usage{
universalPanMammalianClocks(
  betaM,
  speciesName,
  anageData = NULL,
  minCoverage = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{betaM}{A numeric matrix (Rows: CpGs, Cols: Samples).}

\item{speciesName}{A character string or vector specifying the Latin species
name(s) (e.g., "Homo sapiens").}

\item{anageData}{A data.frame containing the AnAge database information.
Must include 'SpeciesLatinName', 'GestationTimeInYears',
'averagedMaturity.yrs', and 'maxAge'.}

\item{minCoverage}{Numeric (0-1). Minimum required proportion of CpGs present.
Default is 0.5.}

\item{verbose}{Logical. Whether to print status messages.}
}
\value{
A data.frame containing the 'Sample', 'SpeciesLatinName',
and the calculated ages: 'DNAmAgePanMammalianClock1','DNAmRelativeAge',
'DNAmAgePanMammalianClock2', 'DNAmRelativeAdultAge' and
'DNAmAgePanMammalianClock3'. Any additional columns
from the input \code{sampleInfo} (like 'Age', 'Tissue') will also be returned.
}
\description{
Applies the three universal pan-mammalian epigenetic clocks (Clock 1, 2, and 3)
from Lu et al. (2023) to a given dataset of DNA methylation values.
}
\details{
This function encapsulates the complete logic from the original publication's
script. It merges sample metadata with AnAge data, processes methylation
values, calculates the three clocks by matrix multiplication with pre-trained
models, and performs the necessary inverse mathematical transformations
to report age in years.
}
\examples{
tursiopsExample <- loadOmniAgeRdata(
    "omniager_tursiops_example",
    verbose = FALSE
)

## This anage_data is from
## https://github.com/shorvath/MammalianMethylationConsortium
anageData <- loadOmniAgeRdata(
    "omniager_anage_data",
    verbose = FALSE
)
# Run the calculation with progress messages
clockResults <- universalPanMammalianClocks(
    betaM = tursiopsExample$beta_m,
    speciesName = tursiopsExample$PhenoTypes$SpeciesLatinName,
    anageData = anageData
)

}
\references{
Lu, A.T., Fei, Z., Haghani, A. et al.
Universal DNA methylation age across mammalian tissues.
\emph{Nat Aging.} 2023
}
