% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weidner.R
\name{weidnerClock}
\alias{weidnerClock}
\title{Calculate Weidner Epigenetic Age (3-CpG Blood Clock)}
\usage{
weidnerClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A \strong{numeric vector} containing the predicted DNAm age
for each sample. The vector is named with the sample IDs from the \code{rownames}
of \code{betaM}.
}
\description{
Estimates biological age using the specific 3-CpG signature described by
Weidner et al. (2014). This model is a multivariate linear regression
originally designed for pyrosequencing data derived from blood samples,
but it can also be applied to microarray data.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
weidnerClockOut <- weidnerClock(hannumBmiqM)
}
\references{
Weidner, C.I., Lin, Q., Koch, C.M. et al.
Aging of blood can be tracked by DNA methylation changes
at just three CpG sites.
\emph{Genome Biol} 2014
}
