% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WuClock.R
\name{WuClock}
\alias{WuClock}
\title{Wu's Epigenetic Clock for Pediatric Age Estimation}
\usage{
WuClock(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of DNA methylation beta values.
\strong{Rows must correspond to samples and columns to CpGs.}
\code{rownames} (sample IDs) and \code{colnames} (CpG probe IDs) are required.
The matrix should not contain \code{NA} values.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{beta.m}.
}
\description{
Estimates biological age (in years) based on DNA methylation data using the method
proposed by Wu et al. (2019). This clock is specifically designed for pediatric
cohorts and utilizes a non-linear transformation to capture rapid developmental
changes in early life.
}
\details{
The calculation involves two main steps:
\enumerate{
\item Calculation of a linear predictor using weighted CpG beta values.
\item Transformation of the linear predictor into biological age (years) using
a specific "anti-transformation" function with a toddler age offset of 48 months.
}

\strong{Data Requirements:}
The input \code{beta.m} must be a matrix of Beta values (0 to 1). The function
expects CpG probes as row names.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
WuClock_o <- WuClock(hannum_bmiq_m)

}
\references{
Wu, Xiaohui et al.
DNA methylation profile is a quantitative measure of biological aging in children
\emph{Aging} 2019
}
