% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WuClock.R
\name{wuClock}
\alias{wuClock}
\title{Wu's Epigenetic Clock for Pediatric Age Estimation}
\usage{
wuClock(betaM, minCoverage = 0.5, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.5.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted biological ages. The vector is
named using the sample IDs from the \code{rownames} of \code{betaM}.
}
\description{
Estimates biological age (in years) based on DNA methylation data using the method
proposed by Wu et al. (2019). This clock is specifically designed for pediatric
cohorts and utilizes a non-linear transformation to capture rapid developmental
changes in early life.
}
\details{
The calculation involves two main steps:
\enumerate{
\item Calculation of a linear predictor using weighted CpG beta values.
\item Transformation of the linear predictor into biological age (years) using
a specific "anti-transformation" function with a toddler age offset of 48 months.
}

\strong{Data Requirements:}
The input \code{beta.m} must be a matrix of Beta values (0 to 1). The function
expects CpG probes as row names.
}
\examples{
downloadOmniAgeRExample("Hannum_example")
loadOmniAgeRExample("Hannum_example")
wuClockOut <- wuClock(hannum_bmiq_m)
}
\references{
Wu, Xiaohui et al.
DNA methylation profile is a quantitative measure of biological aging in children
\emph{Aging} 2019
}
