% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhang10.R
\name{Zhang10}
\alias{Zhang10}
\title{Calculate Zhang10 DNAm Age (2017)}
\usage{
Zhang10(beta.m)
}
\arguments{
\item{beta.m}{A numeric matrix of beta values. Rows should be CpG probes and columns should be individual samples.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to
the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Zhang10 epigenetic clock age (2017) from a DNA methylation beta value matrix.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
Zhang10.out <- Zhang10(hannum_bmiq_m)
}
\references{
Zhang Y, Wilson R, Heiss J, et al.
DNA methylation signatures in peripheral blood strongly predict all-cause mortality.
\emph{Nat Commun.} 2017
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate
multiple clocks simultaneously.
}
