% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhang10.R
\name{zhang10}
\alias{zhang10}
\title{Calculate Zhang10 DNAm Age (2017)}
\usage{
zhang10(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted DNAm ages, with names corresponding to
the sample IDs from the input matrix's column names.
}
\description{
A function to calculate the Zhang10 epigenetic clock age (2017)
from a DNA methylation beta value matrix.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
zhang10Out <- zhang10(hannumBmiqM)

}
\references{
Zhang Y, Wilson R, Heiss J, et al.
DNA methylation signatures in peripheral blood strongly predict
all-cause mortality.
\emph{Nat Commun.} 2017
}
