% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZhangClock.R
\name{ZhangClock}
\alias{ZhangClock}
\title{Predicts DNA methylation age using the Zhang clock(Elastic Net model)
#'}
\usage{
ZhangClock(beta.m)
}
\arguments{
\item{beta.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
A numeric vector of predicted ages, with sample names preserved.
}
\description{
This function takes a matrix of DNA methylation beta values and calculates the epigenetic age for each sample based on the elastic net model developed by Zhang et al. (2019). n.
}
\details{
This function implements the elastic net epigenetic clock from Zhang et al. (2019). A unique feature of this clock is the pre-processing step. Instead of using raw beta values, it first performs a per-sample standardization. Specifically, it calculates a Z-score for each CpG based on the mean and standard deviation of all measured CpGs within that same sample. The final age is then predicted by taking the weighted sum of these standardized values using the model's pre-defined coefficients.
}
\examples{
download_OmniAgeR_example("Hannum_example")
load_OmniAgeR_example("Hannum_example")
ZhangClock.out <- ZhangClock(hannum_bmiq_m)

}
\references{
Zhang Q, Vallerga CL, Walker RM, et al.
Improved precision of epigenetic clock estimates across tissues and its implication for biological ageing.
\emph{Genome Med.} 2019
}
\seealso{
The main function \code{\link{EpiAge}} can be used to calculate multiple clocks simultaneously.
}
