% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZhangClock.R
\name{zhangClock}
\alias{zhangClock}
\title{Predicts DNA methylation age using the Zhang clock(Elastic Net model)
#'}
\usage{
zhangClock(betaM, minCoverage = 0, verbose = TRUE)
}
\arguments{
\item{betaM}{A numeric matrix of DNA methylation beta values.
\code{rownames} (CpG probe IDs) and \code{colnames} (Sample IDs) are required.
The matrix should not contain \code{NA} values.}

\item{minCoverage}{A numeric value (0-1). The minimum proportion of
required CpGs that must be present. Default is 0.}

\item{verbose}{A logical flag. If \code{TRUE} (default), prints status messages.}
}
\value{
A numeric vector of predicted ages, with sample names preserved.
}
\description{
This function takes a matrix of DNA methylation beta values and calculates
the epigenetic age for each sample based on the elastic net model developed
by Zhang et al. (2019).
}
\details{
This function implements the elastic net epigenetic clock from
Zhang et al. (2019). A unique feature of this clock is the
pre-processing step. Instead of using raw beta values, it
first performs a per-sample standardization. Specifically, it calculates a
Z-score for each CpG based on the mean and standard deviation of all
measured CpGs within that same sample. The final age is then predicted
by taking the weighted sum of these standardized values using the model's
pre-defined coefficients.
}
\examples{
hannumBmiqM <- loadOmniAgeRdata(
    "omniager_hannum_example",
    verbose = FALSE
)[[1]]
zhangClockOut <- zhangClock(hannumBmiqM)
}
\references{
Zhang Q, Vallerga CL, Walker RM, et al.
Improved precision of epigenetic clock estimates across tissues and
its implication for biological ageing.
\emph{Genome Med.} 2019
}
